@@ -19,14 +19,37 @@ classifiers = [
     "Programming Language :: Python :: 3 :: Only",
     "Topic :: Scientific/Engineering :: GIS",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
-    "fiona >= 1.8.19",
+    "numpy >= 1.22",
+    "pyogrio >= 0.7.2",
     "packaging",
-    "pandas >= 1.1.0",
-    "pyproj >= 3.0.1",
-    "shapely >= 1.7.1",
+    "pandas >= 1.4.0",
+    "pyproj >= 3.3.0",
+    "shapely >= 2.0.0",
 ]
+
+[project.optional-dependencies]
+all = [
+    "psycopg-binary>=3.1.0",
+    "SQLAlchemy>=1.3",
+    "geopy",
+    "matplotlib>=3.5.0",
+    "mapclassify",
+    "xyzservices",
+    "folium",
+    "GeoAlchemy2",
+    "pyarrow>=8.0.0",
+]
+dev = [
+    "pytest>=3.1.0",
+    "pytest-cov",
+    "pytest-xdist",
+    "codecov",
+    "black",
+    "pre-commit",
+]
+
 
 [project.readme]
 text = """\
