@@ -1,12 +1,12 @@
-from geopandas import GeoDataFrame, GeoSeries, read_file, datasets
+from geopandas import GeoDataFrame, GeoSeries, read_file
 import numpy as np
 import pandas as pd
 from shapely.geometry import Point
-
+from geopandas.tests.util import _NYBB
 
 class CRS:
     def setup(self):
-        nybb = read_file(datasets.get_path("nybb"))
+        nybb = read_file(_NYBB)
         self.long_nybb = GeoDataFrame(pd.concat(10 * [nybb]), crs=nybb.crs)
 
         num_points = 20000
