@@ -28,14 +28,14 @@ axs[0, 0].set_title("s.buffer(0.2)")
 s.buffer(0.2, resolution=2).plot(ax=axs[0, 1], alpha=0.6)
 axs[0, 1].set_title("s.buffer(0.2, resolution=2)")
 
-s.buffer(0.2, cap_style=2).plot(ax=axs[1, 0], alpha=0.6)
-axs[1, 0].set_title("s.buffer(0.2, cap_style=2)")
+s.buffer(0.2, cap_style="square").plot(ax=axs[1, 0], alpha=0.6)
+axs[1, 0].set_title('s.buffer(0.2, cap_style="square")')
 
-s.buffer(0.2, cap_style=3).plot(ax=axs[1, 1], alpha=0.6)
-axs[1, 1].set_title("s.buffer(0.2, cap_style=3)")
+s.buffer(0.2, cap_style="flat").plot(ax=axs[1, 1], alpha=0.6)
+axs[1, 1].set_title('s.buffer(0.2, cap_style="flat")')
 
-s.buffer(0.2, join_style=2).plot(ax=axs[2, 0], alpha=0.6)
-axs[2, 0].set_title("s.buffer(0.2, join_style=2)")
+s.buffer(0.2, join_style="mitre").plot(ax=axs[2, 0], alpha=0.6)
+axs[2, 0].set_title('s.buffer(0.2, join_style="mitre")')
 
-s.buffer(0.2, join_style=3).plot(ax=axs[2, 1], alpha=0.6)
-axs[2, 1].set_title("s.buffer(0.2, join_style=3)")
+s.buffer(0.2, join_style="bevel").plot(ax=axs[2, 1], alpha=0.6)
+axs[2, 1].set_title('s.buffer(0.2, join_style="bevel")')
