@@ -1,7 +1,7 @@
+import inspect
 import numbers
 import operator
 import warnings
-import inspect
 from functools import lru_cache
 
 import numpy as np
@@ -15,21 +15,35 @@ from pandas.api.extensions import (
 import shapely
 import shapely.affinity
 import shapely.geometry
-from shapely.geometry.base import BaseGeometry
 import shapely.ops
 import shapely.wkt
-from pyproj import CRS, Transformer
+from shapely.geometry.base import BaseGeometry
 
-try:
-    import pygeos
-except ImportError:
-    geos = None
+from ._compat import HAS_PYPROJ, requires_pyproj
+from .sindex import SpatialIndex
 
-from . import _compat as compat
-from . import _vectorized as vectorized
-from .sindex import _get_sindex_class
+if HAS_PYPROJ:
+    from pyproj import Transformer
 
-TransformerFromCRS = lru_cache(Transformer.from_crs)
+    TransformerFromCRS = lru_cache(Transformer.from_crs)
+
+_names = {
+    "MISSING": None,
+    "NAG": None,
+    "POINT": "Point",
+    "LINESTRING": "LineString",
+    "LINEARRING": "LinearRing",
+    "POLYGON": "Polygon",
+    "MULTIPOINT": "MultiPoint",
+    "MULTILINESTRING": "MultiLineString",
+    "MULTIPOLYGON": "MultiPolygon",
+    "GEOMETRYCOLLECTION": "GeometryCollection",
+}
+
+
+type_mapping = {p.value: _names[p.name] for p in shapely.GeometryType}
+geometry_type_ids = list(type_mapping.keys())
+geometry_type_values = np.array(list(type_mapping.values()), dtype=object)
 
 
 class GeometryDtype(ExtensionDtype):
@@ -102,40 +116,30 @@ def _crs_mismatch_warn(left, right, stacklevel=3):
     )
 
 
+def isna(value):
+    """
+    Check if scalar value is NA-like (None, np.nan or pd.NA).
+
+    Custom version that only works for scalars (returning True or False),
+    as `pd.isna` also works for array-like input returning a boolean array.
+    """
+    if value is None:
+        return True
+    elif isinstance(value, float) and np.isnan(value):
+        return True
+    elif value is pd.NA:
+        return True
+    else:
+        return False
+
+
 # -----------------------------------------------------------------------------
 # Constructors / converters to other formats
 # -----------------------------------------------------------------------------
 
 
-def _geom_to_shapely(geom):
-    """
-    Convert internal representation (PyGEOS or Shapely) to external Shapely object.
-    """
-    if compat.USE_SHAPELY_20:
-        return geom
-    elif not compat.USE_PYGEOS:
-        return geom
-    else:
-        return vectorized._pygeos_to_shapely(geom)
-
-
-def _shapely_to_geom(geom):
-    """
-    Convert external Shapely object to internal representation (PyGEOS or Shapely).
-    """
-    if compat.USE_SHAPELY_20:
-        return geom
-    elif not compat.USE_PYGEOS:
-        return geom
-    else:
-        return vectorized._shapely_to_pygeos(geom)
-
-
 def _is_scalar_geometry(geom):
-    if compat.USE_PYGEOS:
-        return isinstance(geom, (pygeos.Geometry, BaseGeometry))
-    else:
-        return isinstance(geom, BaseGeometry)
+    return isinstance(geom, BaseGeometry)
 
 
 def from_shapely(data, crs=None):
@@ -154,7 +158,30 @@ def from_shapely(data, crs=None):
         such as an authority string (eg "EPSG:4326") or a WKT string.
 
     """
-    return GeometryArray(vectorized.from_shapely(data), crs=crs)
+    if not isinstance(data, np.ndarray):
+        arr = np.empty(len(data), dtype=object)
+        arr[:] = data
+    else:
+        arr = data
+
+    if not shapely.is_valid_input(arr).all():
+        out = []
+
+        for geom in data:
+            if isinstance(geom, BaseGeometry):
+                out.append(geom)
+            elif hasattr(geom, "__geo_interface__"):
+                geom = shapely.geometry.shape(geom)
+                out.append(geom)
+            elif isna(geom):
+                out.append(None)
+            else:
+                raise TypeError(
+                    "Input must be valid geometry objects: {0}".format(geom)
+                )
+        arr = np.array(out, dtype=object)
+
+    return GeometryArray(arr, crs=crs)
 
 
 def to_shapely(geoms):
@@ -163,10 +190,10 @@ def to_shapely(geoms):
     """
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
-    return vectorized.to_shapely(geoms._data)
+    return geoms._data
 
 
-def from_wkb(data, crs=None):
+def from_wkb(data, crs=None, on_invalid="raise"):
     """
     Convert a list or array of WKB objects to a GeometryArray.
 
@@ -178,9 +205,14 @@ def from_wkb(data, crs=None):
         Coordinate Reference System of the geometry objects. Can be anything accepted by
         :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
         such as an authority string (eg "EPSG:4326") or a WKT string.
+    on_invalid: {"raise", "warn", "ignore"}, default "raise"
+        - raise: an exception will be raised if a WKB input geometry is invalid.
+        - warn: a warning will be raised and invalid WKB geometries will be returned as
+          None.
+        - ignore: invalid WKB geometries will be returned as None without a warning.
 
     """
-    return GeometryArray(vectorized.from_wkb(data), crs=crs)
+    return GeometryArray(shapely.from_wkb(data, on_invalid=on_invalid), crs=crs)
 
 
 def to_wkb(geoms, hex=False, **kwargs):
@@ -189,10 +221,10 @@ def to_wkb(geoms, hex=False, **kwargs):
     """
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
-    return vectorized.to_wkb(geoms._data, hex=hex, **kwargs)
+    return shapely.to_wkb(geoms, hex=hex, **kwargs)
 
 
-def from_wkt(data, crs=None):
+def from_wkt(data, crs=None, on_invalid="raise"):
     """
     Convert a list or array of WKT objects to a GeometryArray.
 
@@ -204,9 +236,14 @@ def from_wkt(data, crs=None):
         Coordinate Reference System of the geometry objects. Can be anything accepted by
         :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
         such as an authority string (eg "EPSG:4326") or a WKT string.
+    on_invalid : {"raise", "warn", "ignore"}, default "raise"
+        - raise: an exception will be raised if a WKT input geometry is invalid.
+        - warn: a warning will be raised and invalid WKT geometries will be
+          returned as ``None``.
+        - ignore: invalid WKT geometries will be returned as ``None`` without a warning.
 
     """
-    return GeometryArray(vectorized.from_wkt(data), crs=crs)
+    return GeometryArray(shapely.from_wkt(data, on_invalid=on_invalid), crs=crs)
 
 
 def to_wkt(geoms, **kwargs):
@@ -215,7 +252,7 @@ def to_wkt(geoms, **kwargs):
     """
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
-    return vectorized.to_wkt(geoms._data, **kwargs)
+    return shapely.to_wkt(geoms, **kwargs)
 
 
 def points_from_xy(x, y, z=None, crs=None):
@@ -261,7 +298,12 @@ def points_from_xy(x, y, z=None, crs=None):
     -------
     output : GeometryArray
     """
-    return GeometryArray(vectorized.points_from_xy(x, y, z), crs=crs)
+    x = np.asarray(x, dtype="float64")
+    y = np.asarray(y, dtype="float64")
+    if z is not None:
+        z = np.asarray(z, dtype="float64")
+
+    return GeometryArray(shapely.points(x, y, z), crs=crs)
 
 
 class GeometryArray(ExtensionArray):
@@ -293,25 +335,9 @@ class GeometryArray(ExtensionArray):
         self._sindex = None
 
     @property
-    def data(self):
-        warnings.warn(
-            "Accessing the underlying geometries through the `.data` attribute is "
-            "deprecated and will be removed in GeoPandas 1.0. You can use "
-            "`np.asarray(..)` or the `to_numpy()` method instead.\n"
-            "Note that if you are using PyGEOS and using this attribute to get an "
-            "array of PyGEOS geometries, those other methods will always return an "
-            "array of Shapely geometries. Accessing the underlying PyGEOS geometries "
-            "directly is deprecated, and you should migrate to use Shapely >= 2.0 "
-            "instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return self._data
-
-    @property
     def sindex(self):
         if self._sindex is None:
-            self._sindex = _get_sindex_class()(self._data)
+            self._sindex = SpatialIndex(self._data)
         return self._sindex
 
     @property
@@ -355,7 +381,21 @@ class GeometryArray(ExtensionArray):
     @crs.setter
     def crs(self, value):
         """Sets the value of the crs"""
-        self._crs = None if not value else CRS.from_user_input(value)
+        if HAS_PYPROJ:
+            from pyproj import CRS
+
+            self._crs = None if not value else CRS.from_user_input(value)
+        else:
+            if value is not None:
+                warnings.warn(
+                    "Cannot set the CRS, falling back to None. The CRS support requires"
+                    " the 'pyproj' package, but it is not installed or does not import"
+                    " correctly. The functions depending on CRS will raise an error or"
+                    " may produce unexpected results.",
+                    UserWarning,
+                    stacklevel=2,
+                )
+            self._crs = None
 
     def check_geographic_crs(self, stacklevel):
         """Check CRS and warn if the planar operation is done in a geographic CRS"""
@@ -379,7 +419,7 @@ class GeometryArray(ExtensionArray):
 
     def __getitem__(self, idx):
         if isinstance(idx, numbers.Integral):
-            return _geom_to_shapely(self._data[idx])
+            return self._data[idx]
         # array-like, slice
         # validate and convert IntegerArray/BooleanArray
         # to numpy array, pass-through non-array-like indexers
@@ -400,8 +440,8 @@ class GeometryArray(ExtensionArray):
             if isinstance(key, numbers.Integral):
                 raise ValueError("cannot set a single element with an array")
             self._data[key] = value._data
-        elif isinstance(value, BaseGeometry) or vectorized.isna(value):
-            if vectorized.isna(value):
+        elif isinstance(value, BaseGeometry) or isna(value):
+            if isna(value):
                 # internally only use None as missing value indicator
                 # but accept others
                 value = None
@@ -411,8 +451,7 @@ class GeometryArray(ExtensionArray):
                 raise TypeError("should be valid geometry")
             if isinstance(key, (slice, list, np.ndarray)):
                 value_array = np.empty(1, dtype=object)
-                with compat.ignore_shapely2_warnings():
-                    value_array[:] = [value]
+                value_array[:] = [value]
                 self._data[key] = value_array
             else:
                 self._data[key] = value
@@ -433,17 +472,12 @@ class GeometryArray(ExtensionArray):
         #         )
 
     def __getstate__(self):
-        if compat.USE_SHAPELY_20:
-            return (shapely.to_wkb(self._data), self._crs)
-        elif compat.USE_PYGEOS:
-            return (pygeos.to_wkb(self._data), self._crs)
-        else:
-            return self.__dict__
+        return (shapely.to_wkb(self._data), self._crs)
 
     def __setstate__(self, state):
         if not isinstance(state, dict):
             # pickle file saved with pygeos
-            geoms = vectorized.from_wkb(state[0])
+            geoms = shapely.from_wkb(state[0])
             self._crs = state[1]
             self._sindex = None  # pygeos.STRtree could not be pickled yet
             self._data = geoms
@@ -451,8 +485,6 @@ class GeometryArray(ExtensionArray):
         else:
             if "data" in state:
                 state["_data"] = state.pop("data")
-            if compat.USE_PYGEOS:
-                state["_data"] = vectorized.from_shapely(state["_data"])
             if "_crs" not in state:
                 state["_crs"] = None
             self.__dict__.update(state)
@@ -463,41 +495,64 @@ class GeometryArray(ExtensionArray):
 
     @property
     def is_valid(self):
-        return vectorized.is_valid(self._data)
+        return shapely.is_valid(self._data)
+
+    def is_valid_reason(self):
+        return shapely.is_valid_reason(self._data)
 
     @property
     def is_empty(self):
-        return vectorized.is_empty(self._data)
+        return shapely.is_empty(self._data)
 
     @property
     def is_simple(self):
-        return vectorized.is_simple(self._data)
+        return shapely.is_simple(self._data)
 
     @property
     def is_ring(self):
-        return vectorized.is_ring(self._data)
+        return shapely.is_ring(self._data)
 
     @property
     def is_closed(self):
-        return vectorized.is_closed(self._data)
+        return shapely.is_closed(self._data)
+
+    @property
+    def is_ccw(self):
+        return shapely.is_ccw(self._data)
 
     @property
     def has_z(self):
-        return vectorized.has_z(self._data)
+        return shapely.has_z(self._data)
 
     @property
     def geom_type(self):
-        return vectorized.geom_type(self._data)
+        res = shapely.get_type_id(self._data)
+        return geometry_type_values[np.searchsorted(geometry_type_ids, res)]
 
     @property
     def area(self):
         self.check_geographic_crs(stacklevel=5)
-        return vectorized.area(self._data)
+        return shapely.area(self._data)
 
     @property
     def length(self):
         self.check_geographic_crs(stacklevel=5)
-        return vectorized.length(self._data)
+        return shapely.length(self._data)
+
+    def count_coordinates(self):
+        return shapely.get_num_coordinates(self._data)
+
+    def count_geometries(self):
+        return shapely.get_num_geometries(self._data)
+
+    def count_interior_rings(self):
+        return shapely.get_num_interior_rings(self._data)
+
+    def get_precision(self):
+        return shapely.get_precision(self._data)
+
+    def get_geometry(self, index):
+        return shapely.get_geometry(self._data, index=index)
 
     #
     # Unary operations that return new geometries
@@ -505,40 +560,37 @@ class GeometryArray(ExtensionArray):
 
     @property
     def boundary(self):
-        return GeometryArray(vectorized.boundary(self._data), crs=self.crs)
+        return GeometryArray(shapely.boundary(self._data), crs=self.crs)
 
     @property
     def centroid(self):
         self.check_geographic_crs(stacklevel=5)
-        return GeometryArray(vectorized.centroid(self._data), crs=self.crs)
+        return GeometryArray(shapely.centroid(self._data), crs=self.crs)
 
     def concave_hull(self, ratio, allow_holes):
-        return vectorized.concave_hull(self._data, ratio=ratio, allow_holes=allow_holes)
+        return shapely.concave_hull(self._data, ratio=ratio, allow_holes=allow_holes)
 
     @property
     def convex_hull(self):
-        return GeometryArray(vectorized.convex_hull(self._data), crs=self.crs)
-
-    def delaunay_triangles(self, tolerance, only_edges):
-        return GeometryArray(
-            vectorized.delaunay_triangles(self._data, tolerance, only_edges),
-            crs=self.crs,
-        )
+        return GeometryArray(shapely.convex_hull(self._data), crs=self.crs)
 
     @property
     def envelope(self):
-        return GeometryArray(vectorized.envelope(self._data), crs=self.crs)
+        return GeometryArray(shapely.envelope(self._data), crs=self.crs)
+
+    def minimum_rotated_rectangle(self):
+        return GeometryArray(shapely.oriented_envelope(self._data), crs=self.crs)
 
     @property
     def exterior(self):
-        return GeometryArray(vectorized.exterior(self._data), crs=self.crs)
+        return GeometryArray(shapely.get_exterior_ring(self._data), crs=self.crs)
 
     def extract_unique_points(self):
-        return GeometryArray(vectorized.extract_unique_points(self._data), crs=self.crs)
+        return GeometryArray(shapely.extract_unique_points(self._data), crs=self.crs)
 
     def offset_curve(self, distance, quad_segs=8, join_style="round", mitre_limit=5.0):
         return GeometryArray(
-            vectorized.offset_curve(
+            shapely.offset_curve(
                 self._data,
                 distance,
                 quad_segs=quad_segs,
@@ -551,28 +603,80 @@ class GeometryArray(ExtensionArray):
     @property
     def interiors(self):
         # no GeometryArray as result
-        return vectorized.interiors(self._data)
+        has_non_poly = False
+        inner_rings = []
+        for geom in self._data:
+            interior_ring_seq = getattr(geom, "interiors", None)
+            # polygon case
+            if interior_ring_seq is not None:
+                inner_rings.append(list(interior_ring_seq))
+            # non-polygon case
+            else:
+                has_non_poly = True
+                inner_rings.append(None)
+        if has_non_poly:
+            warnings.warn(
+                "Only Polygon objects have interior rings. For other "
+                "geometry types, None is returned.",
+                stacklevel=2,
+            )
+        # need to allocate empty first in case of all empty lists in inner_rings
+        data = np.empty(len(inner_rings), dtype=object)
+        data[:] = inner_rings
+        return data
 
-    def representative_point(self):
-        return GeometryArray(vectorized.representative_point(self._data), crs=self.crs)
-
-    def minimum_bounding_circle(self):
+    def remove_repeated_points(self, tolerance=0.0):
         return GeometryArray(
-            vectorized.minimum_bounding_circle(self._data), crs=self.crs
+            shapely.remove_repeated_points(self._data, tolerance=tolerance),
+            crs=self.crs,
         )
 
+    def representative_point(self):
+        return GeometryArray(shapely.point_on_surface(self._data), crs=self.crs)
+
+    def minimum_bounding_circle(self):
+        return GeometryArray(shapely.minimum_bounding_circle(self._data), crs=self.crs)
+
     def minimum_bounding_radius(self):
-        return vectorized.minimum_bounding_radius(self._data)
+        return shapely.minimum_bounding_radius(self._data)
+
+    def minimum_clearance(self):
+        return shapely.minimum_clearance(self._data)
 
     def normalize(self):
-        return GeometryArray(vectorized.normalize(self._data), crs=self.crs)
+        return GeometryArray(shapely.normalize(self._data), crs=self.crs)
 
     def make_valid(self):
-        return GeometryArray(vectorized.make_valid(self._data), crs=self.crs)
+        return GeometryArray(shapely.make_valid(self._data), crs=self.crs)
+
+    def reverse(self):
+        return GeometryArray(shapely.reverse(self._data), crs=self.crs)
 
     def segmentize(self, max_segment_length):
         return GeometryArray(
-            vectorized.segmentize(self._data, max_segment_length),
+            shapely.segmentize(self._data, max_segment_length),
+            crs=self.crs,
+        )
+
+    def force_2d(self):
+        return GeometryArray(shapely.force_2d(self._data), crs=self.crs)
+
+    def force_3d(self, z=0):
+        return GeometryArray(shapely.force_3d(self._data, z=z), crs=self.crs)
+
+    def transform(self, transformation, include_z=False):
+        return GeometryArray(
+            shapely.transform(self._data, transformation, include_z), crs=self.crs
+        )
+
+    def line_merge(self, directed=False):
+        return GeometryArray(
+            shapely.line_merge(self._data, directed=directed), crs=self.crs
+        )
+
+    def set_precision(self, grid_size, mode="valid_output"):
+        return GeometryArray(
+            shapely.set_precision(self._data, grid_size=grid_size, mode=mode),
             crs=self.crs,
         )
 
@@ -592,7 +696,7 @@ class GeometryArray(ExtensionArray):
                 _crs_mismatch_warn(left, right, stacklevel=7)
             right = right._data
 
-        return getattr(vectorized, op)(left._data, right, **kwargs)
+        return getattr(shapely, op)(left._data, right, **kwargs)
 
     def covers(self, other):
         return self._binary_method("covers", self, other)
@@ -602,6 +706,9 @@ class GeometryArray(ExtensionArray):
 
     def contains(self, other):
         return self._binary_method("contains", self, other)
+
+    def contains_properly(self, other):
+        return self._binary_method("contains_properly", self, other)
 
     def crosses(self, other):
         return self._binary_method("crosses", self, other)
@@ -624,12 +731,21 @@ class GeometryArray(ExtensionArray):
     def within(self, other):
         return self._binary_method("within", self, other)
 
+    def dwithin(self, other, distance):
+        self.check_geographic_crs(stacklevel=6)
+        return self._binary_method("dwithin", self, other, distance=distance)
+
     def geom_equals_exact(self, other, tolerance):
         return self._binary_method("equals_exact", self, other, tolerance=tolerance)
 
     def geom_almost_equals(self, other, decimal):
+        warnings.warn(
+            "The 'geom_almost_equals()' method is deprecated because the name is "
+            "confusing. The 'geom_equals_exact()' method should be used instead.",
+            FutureWarning,
+            stacklevel=2,
+        )
         return self.geom_equals_exact(other, 0.5 * 10 ** (-decimal))
-        # return _binary_predicate("almost_equals", self, other, decimal=decimal)
 
     #
     # Binary operations that return new geometries
@@ -637,7 +753,7 @@ class GeometryArray(ExtensionArray):
 
     def clip_by_rect(self, xmin, ymin, xmax, ymax):
         return GeometryArray(
-            vectorized.clip_by_rect(self._data, xmin, ymin, xmax, ymax), crs=self.crs
+            shapely.clip_by_rect(self._data, xmin, ymin, xmax, ymax), crs=self.crs
         )
 
     def difference(self, other):
@@ -658,6 +774,21 @@ class GeometryArray(ExtensionArray):
     def union(self, other):
         return GeometryArray(self._binary_method("union", self, other), crs=self.crs)
 
+    def shortest_line(self, other):
+        return GeometryArray(
+            self._binary_method("shortest_line", self, other), crs=self.crs
+        )
+
+    def snap(self, other, tolerance):
+        return GeometryArray(
+            self._binary_method("snap", self, other, tolerance=tolerance), crs=self.crs
+        )
+
+    def shared_paths(self, other):
+        return GeometryArray(
+            self._binary_method("shared_paths", self, other), crs=self.crs
+        )
+
     #
     # Other operations
     #
@@ -670,67 +801,110 @@ class GeometryArray(ExtensionArray):
         self.check_geographic_crs(stacklevel=6)
         return self._binary_method("hausdorff_distance", self, other, **kwargs)
 
+    def frechet_distance(self, other, **kwargs):
+        self.check_geographic_crs(stacklevel=6)
+        return self._binary_method("frechet_distance", self, other, **kwargs)
+
     def buffer(self, distance, resolution=16, **kwargs):
         if not (isinstance(distance, (int, float)) and distance == 0):
             self.check_geographic_crs(stacklevel=5)
         return GeometryArray(
-            vectorized.buffer(self._data, distance, resolution=resolution, **kwargs),
+            shapely.buffer(self._data, distance, quad_segs=resolution, **kwargs),
             crs=self.crs,
         )
 
     def interpolate(self, distance, normalized=False):
         self.check_geographic_crs(stacklevel=5)
         return GeometryArray(
-            vectorized.interpolate(self._data, distance, normalized=normalized),
+            shapely.line_interpolate_point(self._data, distance, normalized=normalized),
             crs=self.crs,
         )
 
     def simplify(self, tolerance, preserve_topology=True):
         return GeometryArray(
-            vectorized.simplify(
+            shapely.simplify(
                 self._data, tolerance, preserve_topology=preserve_topology
             ),
             crs=self.crs,
         )
 
     def project(self, other, normalized=False):
-        if isinstance(other, BaseGeometry):
-            other = _shapely_to_geom(other)
-        elif isinstance(other, GeometryArray):
+        if isinstance(other, GeometryArray):
             other = other._data
-        return vectorized.project(self._data, other, normalized=normalized)
+        return shapely.line_locate_point(self._data, other, normalized=normalized)
 
     def relate(self, other):
         if isinstance(other, GeometryArray):
             other = other._data
-        return vectorized.relate(self._data, other)
+        return shapely.relate(self._data, other)
+
+    def relate_pattern(self, other, pattern):
+        if isinstance(other, GeometryArray):
+            other = other._data
+        return shapely.relate_pattern(self._data, other, pattern)
 
     #
     # Reduction operations that return a Shapely geometry
     #
 
     def unary_union(self):
-        return vectorized.unary_union(self._data)
+        warnings.warn(
+            "The 'unary_union' attribute is deprecated, "
+            "use the 'union_all' method instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return self.union_all()
+
+    def union_all(self, method="unary"):
+        if method == "coverage":
+            return shapely.coverage_union_all(self._data)
+        elif method == "unary":
+            return shapely.union_all(self._data)
+        else:
+            raise ValueError(
+                f"Method '{method}' not recognized. Use 'coverage' or 'unary'."
+            )
+
+    def intersection_all(self):
+        return shapely.intersection_all(self._data)
 
     #
     # Affinity operations
     #
 
+    @staticmethod
+    def _affinity_method(op, left, *args, **kwargs):
+        # not all shapely.affinity methods can handle empty geometries:
+        # affine_transform itself works (as well as translate), but rotate, scale
+        # and skew fail (they try to unpack the bounds).
+        # Here: consistently returning empty geom for input empty geom
+        out = []
+        for geom in left:
+            if geom is None or geom.is_empty:
+                res = geom
+            else:
+                res = getattr(shapely.affinity, op)(geom, *args, **kwargs)
+            out.append(res)
+        data = np.empty(len(left), dtype=object)
+        data[:] = out
+        return data
+
     def affine_transform(self, matrix):
         return GeometryArray(
-            vectorized._affinity_method("affine_transform", self._data, matrix),
+            self._affinity_method("affine_transform", self._data, matrix),
             crs=self.crs,
         )
 
     def translate(self, xoff=0.0, yoff=0.0, zoff=0.0):
         return GeometryArray(
-            vectorized._affinity_method("translate", self._data, xoff, yoff, zoff),
+            self._affinity_method("translate", self._data, xoff, yoff, zoff),
             crs=self.crs,
         )
 
     def rotate(self, angle, origin="center", use_radians=False):
         return GeometryArray(
-            vectorized._affinity_method(
+            self._affinity_method(
                 "rotate", self._data, angle, origin=origin, use_radians=use_radians
             ),
             crs=self.crs,
@@ -738,7 +912,7 @@ class GeometryArray(ExtensionArray):
 
     def scale(self, xfact=1.0, yfact=1.0, zfact=1.0, origin="center"):
         return GeometryArray(
-            vectorized._affinity_method(
+            self._affinity_method(
                 "scale", self._data, xfact, yfact, zfact, origin=origin
             ),
             crs=self.crs,
@@ -746,12 +920,13 @@ class GeometryArray(ExtensionArray):
 
     def skew(self, xs=0.0, ys=0.0, origin="center", use_radians=False):
         return GeometryArray(
-            vectorized._affinity_method(
+            self._affinity_method(
                 "skew", self._data, xs, ys, origin=origin, use_radians=use_radians
             ),
             crs=self.crs,
         )
 
+    @requires_pyproj
     def to_crs(self, crs=None, epsg=None):
         """Returns a ``GeometryArray`` with all geometries transformed to a new
         coordinate reference system.
@@ -821,6 +996,8 @@ class GeometryArray(ExtensionArray):
         - Prime Meridian: Greenwich
 
         """
+        from pyproj import CRS
+
         if self.crs is None:
             raise ValueError(
                 "Cannot transform naive geometries.  "
@@ -839,9 +1016,10 @@ class GeometryArray(ExtensionArray):
 
         transformer = TransformerFromCRS(self.crs, crs, always_xy=True)
 
-        new_data = vectorized.transform(self._data, transformer.transform)
+        new_data = transform(self._data, transformer.transform)
         return GeometryArray(new_data, crs=crs)
 
+    @requires_pyproj
     def estimate_utm_crs(self, datum_name="WGS 84"):
         """Returns the estimated UTM CRS based on the bounds of the dataset.
 
@@ -880,11 +1058,9 @@ class GeometryArray(ExtensionArray):
         - Ellipsoid: WGS 84
         - Prime Meridian: Greenwich
         """
-        try:
-            from pyproj.aoi import AreaOfInterest
-            from pyproj.database import query_utm_crs_info
-        except ImportError:
-            raise RuntimeError("pyproj 3+ required for estimate_utm_crs.")
+        from pyproj import CRS
+        from pyproj.aoi import AreaOfInterest
+        from pyproj.database import query_utm_crs_info
 
         if not self.crs:
             raise RuntimeError("crs must be set to estimate UTM CRS.")
@@ -896,28 +1072,21 @@ class GeometryArray(ExtensionArray):
         # ensure using geographic coordinates
         else:
             transformer = TransformerFromCRS(self.crs, "EPSG:4326", always_xy=True)
-            if compat.PYPROJ_GE_31:
-                minx, miny, maxx, maxy = transformer.transform_bounds(
-                    minx, miny, maxx, maxy
-                )
-                y_center = np.mean([miny, maxy])
-                # crossed the antimeridian
-                if minx > maxx:
-                    # shift maxx from [-180,180] to [0,360]
-                    # so both numbers are positive for center calculation
-                    # Example: -175 to 185
-                    maxx += 360
-                    x_center = np.mean([minx, maxx])
-                    # shift back to [-180,180]
-                    x_center = ((x_center + 180) % 360) - 180
-                else:
-                    x_center = np.mean([minx, maxx])
+            minx, miny, maxx, maxy = transformer.transform_bounds(
+                minx, miny, maxx, maxy
+            )
+            y_center = np.mean([miny, maxy])
+            # crossed the antimeridian
+            if minx > maxx:
+                # shift maxx from [-180,180] to [0,360]
+                # so both numbers are positive for center calculation
+                # Example: -175 to 185
+                maxx += 360
+                x_center = np.mean([minx, maxx])
+                # shift back to [-180,180]
+                x_center = ((x_center + 180) % 360) - 180
             else:
-                lon, lat = transformer.transform(
-                    (minx, maxx, minx, maxx), (miny, miny, maxy, maxy)
-                )
-                x_center = np.mean(lon)
-                y_center = np.mean(lat)
+                x_center = np.mean([minx, maxx])
 
         utm_crs_list = query_utm_crs_info(
             datum_name=datum_name,
@@ -945,10 +1114,10 @@ class GeometryArray(ExtensionArray):
             if empty.any():
                 nonempty = ~empty
                 coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
-                coords[nonempty] = vectorized.get_x(self._data[nonempty])
+                coords[nonempty] = shapely.get_x(self._data[nonempty])
                 return coords
             else:
-                return vectorized.get_x(self._data)
+                return shapely.get_x(self._data)
         else:
             message = "x attribute access only provided for Point geometries"
             raise ValueError(message)
@@ -961,10 +1130,10 @@ class GeometryArray(ExtensionArray):
             if empty.any():
                 nonempty = ~empty
                 coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
-                coords[nonempty] = vectorized.get_y(self._data[nonempty])
+                coords[nonempty] = shapely.get_y(self._data[nonempty])
                 return coords
             else:
-                return vectorized.get_y(self._data)
+                return shapely.get_y(self._data)
         else:
             message = "y attribute access only provided for Point geometries"
             raise ValueError(message)
@@ -977,17 +1146,17 @@ class GeometryArray(ExtensionArray):
             if empty.any():
                 nonempty = ~empty
                 coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
-                coords[nonempty] = vectorized.get_z(self._data[nonempty])
+                coords[nonempty] = shapely.get_z(self._data[nonempty])
                 return coords
             else:
-                return vectorized.get_z(self._data)
+                return shapely.get_z(self._data)
         else:
             message = "z attribute access only provided for Point geometries"
             raise ValueError(message)
 
     @property
     def bounds(self):
-        return vectorized.bounds(self._data)
+        return shapely.bounds(self._data)
 
     @property
     def total_bounds(self):
@@ -996,14 +1165,19 @@ class GeometryArray(ExtensionArray):
             # TODO with numpy >= 1.15, the 'initial' argument can be used
             return np.array([np.nan, np.nan, np.nan, np.nan])
         b = self.bounds
-        return np.array(
-            (
-                np.nanmin(b[:, 0]),  # minx
-                np.nanmin(b[:, 1]),  # miny
-                np.nanmax(b[:, 2]),  # maxx
-                np.nanmax(b[:, 3]),  # maxy
+        with warnings.catch_warnings():
+            # if all rows are empty geometry / none, nan is expected
+            warnings.filterwarnings(
+                "ignore", r"All-NaN slice encountered", RuntimeWarning
             )
-        )
+            return np.array(
+                (
+                    np.nanmin(b[:, 0]),  # minx
+                    np.nanmin(b[:, 1]),  # miny
+                    np.nanmax(b[:, 2]),  # maxx
+                    np.nanmax(b[:, 3]),  # maxy
+                )
+            )
 
     # -------------------------------------------------------------------------
     # general array like compat
@@ -1031,42 +1205,23 @@ class GeometryArray(ExtensionArray):
         if allow_fill:
             if fill_value is None or pd.isna(fill_value):
                 fill_value = None
-            elif isinstance(fill_value, BaseGeometry):
-                fill_value = _shapely_to_geom(fill_value)
             elif not _is_scalar_geometry(fill_value):
                 raise TypeError("provide geometry or None as fill value")
 
         result = take(self._data, indices, allow_fill=allow_fill, fill_value=fill_value)
         if allow_fill and fill_value is None:
-            result[pd.isna(result)] = None
+            result[~shapely.is_valid_input(result)] = None
         return GeometryArray(result, crs=self.crs)
 
-    def _fill(self, idx, value):
-        """
-        Fill index locations with ``value``.
+    # compat for pandas < 3.0
+    def _pad_or_backfill(
+        self, method, limit=None, limit_area=None, copy=True, **kwargs
+    ):
+        return super()._pad_or_backfill(
+            method=method, limit=limit, limit_area=limit_area, copy=copy, **kwargs
+        )
 
-        ``value`` should be a BaseGeometry or a GeometryArray.
-        """
-        if vectorized.isna(value):
-            value = [None]
-        elif _is_scalar_geometry(value):
-            value = [value]
-        elif isinstance(value, GeometryArray):
-            value = value[idx]
-        else:
-            raise TypeError(
-                "'value' parameter must be None, a scalar geometry, or a GeoSeries, "
-                f"but you passed a {type(value).__name__!r}"
-            )
-
-        value_arr = np.empty(len(value), dtype=object)
-        with compat.ignore_shapely2_warnings():
-            value_arr[:] = _shapely_to_geom(value)
-
-        self._data[idx] = value_arr
-        return self
-
-    def fillna(self, value=None, method=None, limit=None):
+    def fillna(self, value=None, method=None, limit=None, copy=True):
         """
         Fill NA values with geometry (or geometries) or using the specified method.
 
@@ -1083,12 +1238,11 @@ class GeometryArray(ExtensionArray):
             backfill / bfill: use NEXT valid observation to fill gap
 
         limit : int, default None
-            If method is specified, this is the maximum number of consecutive
-            NaN values to forward/backward fill. In other words, if there is
-            a gap with more than this number of consecutive NaNs, it will only
-            be partially filled. If method is not specified, this is the
-            maximum number of entries along the entire axis where NaNs will be
-            filled.
+            The maximum number of entries where NA values will be filled.
+
+        copy : bool, default True
+            Whether to make a copy of the data before filling. If False, then
+            the original should be modified and no new memory should be allocated.
 
         Returns
         -------
@@ -1098,8 +1252,34 @@ class GeometryArray(ExtensionArray):
             raise NotImplementedError("fillna with a method is not yet supported")
 
         mask = self.isna()
-        new_values = self.copy()
-        return new_values._fill(mask, value) if mask.any() else new_values
+        if copy:
+            new_values = self.copy()
+        else:
+            new_values = self
+
+        if not mask.any():
+            return new_values
+
+        if limit is not None and limit < len(self):
+            modify = mask.cumsum() > limit
+            if modify.any():
+                mask[modify] = False
+
+        if isna(value):
+            value = [None]
+        elif _is_scalar_geometry(value):
+            value = [value]
+        elif isinstance(value, GeometryArray):
+            value = value[mask]
+        else:
+            raise TypeError(
+                "'value' parameter must be None, a scalar geometry, or a GeoSeries, "
+                f"but you passed a {type(value).__name__!r}"
+            )
+        value_arr = np.asarray(value, dtype=object)
+
+        new_values._data[mask] = value_arr
+        return new_values
 
     def astype(self, dtype, copy=True):
         """
@@ -1134,18 +1314,19 @@ class GeometryArray(ExtensionArray):
                 return pd.array(string_values, dtype=pd_dtype)
             return string_values.astype(dtype, copy=False)
         else:
-            return np.array(self, dtype=dtype, copy=copy)
+            # numpy 2.0 makes copy=False case strict (errors if cannot avoid the copy)
+            # -> in that case use `np.asarray` as backwards compatible alternative
+            # for `copy=None` (when requiring numpy 2+, this can be cleaned up)
+            if not copy:
+                return np.asarray(self, dtype=dtype)
+            else:
+                return np.array(self, dtype=dtype, copy=copy)
 
     def isna(self):
         """
         Boolean NumPy array indicating if each value is missing
         """
-        if compat.USE_SHAPELY_20:
-            return shapely.is_missing(self._data)
-        elif compat.USE_PYGEOS:
-            return pygeos.is_missing(self._data)
-        else:
-            return np.array([g is None for g in self._data], dtype="bool")
+        return shapely.is_missing(self._data)
 
     def value_counts(
         self,
@@ -1167,7 +1348,7 @@ class GeometryArray(ExtensionArray):
         # note ExtensionArray usage of value_counts only specifies dropna,
         # so sort, normalize and bins are not arguments
         values = to_wkb(self)
-        from pandas import Series, Index
+        from pandas import Index, Series
 
         result = Series(values).value_counts(dropna=dropna)
         # value_counts converts None to nan, need to convert back for from_wkb to work
@@ -1258,6 +1439,29 @@ class GeometryArray(ExtensionArray):
         if isinstance(scalars, BaseGeometry):
             scalars = [scalars]
         return from_shapely(scalars)
+
+    @classmethod
+    def _from_sequence_of_strings(cls, strings, *, dtype=None, copy=False):
+        """
+        Construct a new ExtensionArray from a sequence of strings.
+
+        Parameters
+        ----------
+        strings : Sequence
+            Each element will be an instance of the scalar type for this
+            array, ``cls.dtype.type``.
+        dtype : dtype, optional
+            Construct for this particular dtype. This should be a Dtype
+            compatible with the ExtensionArray.
+        copy : bool, default False
+            If True, copy the underlying data.
+
+        Returns
+        -------
+        ExtensionArray
+        """
+        # GH 3099
+        return from_wkt(strings)
 
     def _values_for_factorize(self):
         # type: () -> Tuple[np.ndarray, Any]
@@ -1414,7 +1618,7 @@ class GeometryArray(ExtensionArray):
                         # typically projected coordinates
                         # (in case of unit meter: mm precision)
                         precision = 3
-            return lambda geom: shapely.wkt.dumps(geom, rounding_precision=precision)
+            return lambda geom: shapely.to_wkt(geom, rounding_precision=precision)
         return repr
 
     @classmethod
@@ -1436,15 +1640,14 @@ class GeometryArray(ExtensionArray):
     def _reduce(self, name, skipna=True, **kwargs):
         # including the base class version here (that raises by default)
         # because this was not yet defined in pandas 0.23
-        if name == "any" or name == "all":
-            # TODO(pygeos)
+        if name in ("any", "all"):
             return getattr(to_shapely(self), name)()
         raise TypeError(
             f"'{type(self).__name__}' with dtype {self.dtype} "
             f"does not support reduction '{name}'"
         )
 
-    def __array__(self, dtype=None):
+    def __array__(self, dtype=None, copy=None):
         """
         The numpy array interface.
 
@@ -1452,7 +1655,9 @@ class GeometryArray(ExtensionArray):
         -------
         values : numpy array
         """
-        return to_shapely(self)
+        if copy and (dtype is None or dtype == np.dtype("object")):
+            return self._data.copy()
+        return self._data
 
     def _binop(self, other, op):
         def convert_values(param):
@@ -1491,7 +1696,7 @@ class GeometryArray(ExtensionArray):
         """
         Return for `item in self`.
         """
-        if vectorized.isna(item):
+        if isna(item):
             if (
                 item is self.dtype.na_value
                 or isinstance(item, self.dtype.type)
@@ -1507,15 +1712,20 @@ def _get_common_crs(arr_seq):
     # mask out all None arrays with no crs (most likely auto generated by pandas
     # from concat with missing column)
     arr_seq = [ga for ga in arr_seq if not (ga.isna().all() and ga.crs is None)]
+    # determine unique crs without using a set, because CRS hash can be different
+    # for objects with the same CRS
+    unique_crs = []
+    for arr in arr_seq:
+        if arr.crs not in unique_crs:
+            unique_crs.append(arr.crs)
 
-    crs_set = {arr.crs for arr in arr_seq}
-    crs_not_none = [crs for crs in crs_set if crs is not None]
+    crs_not_none = [crs for crs in unique_crs if crs is not None]
     names = [crs.name for crs in crs_not_none]
 
     if len(crs_not_none) == 0:
         return None
     if len(crs_not_none) == 1:
-        if len(crs_set) != 1:
+        if len(unique_crs) != 1:
             warnings.warn(
                 "CRS not set for some of the concatenation inputs. "
                 f"Setting output's CRS as {names[0]} "
@@ -1528,3 +1738,23 @@ def _get_common_crs(arr_seq):
         f"Cannot determine common CRS for concatenation inputs, got {names}. "
         "Use `to_crs()` to transform geometries to the same CRS before merging."
     )
+
+
+def transform(data, func):
+    has_z = shapely.has_z(data)
+
+    result = np.empty_like(data)
+
+    coords = shapely.get_coordinates(data[~has_z], include_z=False)
+    new_coords_z = func(coords[:, 0], coords[:, 1])
+    result[~has_z] = shapely.set_coordinates(
+        data[~has_z].copy(), np.array(new_coords_z).T
+    )
+
+    coords_z = shapely.get_coordinates(data[has_z], include_z=True)
+    new_coords_z = func(coords_z[:, 0], coords_z[:, 1], coords_z[:, 2])
+    result[has_z] = shapely.set_coordinates(
+        data[has_z].copy(), np.array(new_coords_z).T
+    )
+
+    return result
