@@ -2,34 +2,41 @@ from __future__ import annotations
 
 import json
 import typing
-from typing import Any, Literal
 import warnings
-
+from typing import Any, Literal
 
 import numpy as np
 import pandas as pd
 from pandas import DataFrame, Series
-from pandas.core.accessor import CachedAccessor
 
+import shapely.errors
 from shapely.geometry import mapping, shape
 from shapely.geometry.base import BaseGeometry
 
-from pyproj import CRS
-
+import geopandas
 from geopandas.array import GeometryArray, GeometryDtype, from_shapely, to_wkb, to_wkt
 from geopandas.base import GeoPandasBase, is_geometry_type
-from geopandas.geoseries import GeoSeries
-import geopandas.io
 from geopandas.explore import _explore
-from . import _compat as compat
+from geopandas.geoseries import GeoSeries
+
+from ._compat import HAS_PYPROJ, PANDAS_GE_30
 from ._decorator import doc
+
+if PANDAS_GE_30:
+    from pandas.core.accessor import Accessor
+else:
+    from pandas.core.accessor import CachedAccessor as Accessor
 
 
 if typing.TYPE_CHECKING:
     import os
-    import sqlalchemy.text
+
     import folium
-    from .io.arrow import SUPPORTED_VERSIONS_LITERAL
+    import sqlalchemy.text
+
+    from pyproj import CRS
+
+    from geopandas.io.arrow import SUPPORTED_VERSIONS_LITERAL
 
 
 def _geodataframe_constructor_with_fallback(
@@ -63,7 +70,10 @@ def _ensure_geometry(data, crs: Any | None = None) -> GeoSeries | GeometryArray:
         if data.crs is None and crs is not None:
             # Avoids caching issues/crs sharing issues
             data = data.copy()
-            data.crs = crs
+            if isinstance(data, GeometryArray):
+                data.crs = crs
+            else:
+                data.array.crs = crs
         return data
     else:
         if isinstance(data, Series):
@@ -84,8 +94,8 @@ crs_mismatch_error = (
 
 class GeoDataFrame(GeoPandasBase, DataFrame):
     """
-    A GeoDataFrame object is a pandas.DataFrame that has a column
-    with geometry. In addition to the standard DataFrame constructor arguments,
+    A GeoDataFrame object is a pandas.DataFrame that has one or more columns
+    containing geometry. In addition to the standard DataFrame constructor arguments,
     GeoDataFrame also accepts the following keyword arguments:
 
     Parameters
@@ -94,9 +104,16 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         Coordinate Reference System of the geometry objects. Can be anything accepted by
         :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
         such as an authority string (eg "EPSG:4326") or a WKT string.
-    geometry : str or array (optional)
-        If str, column to use as geometry. If array, will be set as 'geometry'
-        column on GeoDataFrame.
+    geometry : str or array-like (optional)
+        Value to use as the active geometry column.
+        If str, treated as column name to use. If array-like, it will be
+        added as new column named 'geometry' on the GeoDataFrame and set as the
+        active geometry column.
+
+        Note that if ``geometry`` is a (Geo)Series with a
+        name, the name will not be used, a column named "geometry" will still be
+        added. To preserve the name, you can use :meth:`~GeoDataFrame.rename_geometry`
+        to update the geometry column name.
 
     Examples
     --------
@@ -106,9 +123,9 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
     >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
     >>> gdf = geopandas.GeoDataFrame(d, crs="EPSG:4326")
     >>> gdf
-        col1                 geometry
-    0  name1  POINT (1.00000 2.00000)
-    1  name2  POINT (2.00000 1.00000)
+        col1     geometry
+    0  name1  POINT (1 2)
+    1  name2  POINT (2 1)
 
     Notice that the inferred dtype of 'geometry' columns is geometry.
 
@@ -125,9 +142,9 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
     >>> gs = geopandas.GeoSeries.from_wkt(df['wkt'])
     >>> gdf = geopandas.GeoDataFrame(df, geometry=gs, crs="EPSG:4326")
     >>> gdf
-        col1          wkt                 geometry
-    0  name1  POINT (1 2)  POINT (1.00000 2.00000)
-    1  name2  POINT (2 1)  POINT (2.00000 1.00000)
+        col1          wkt     geometry
+    0  name1  POINT (1 2)  POINT (1 2)
+    1  name2  POINT (2 1)  POINT (2 1)
 
     See also
     --------
@@ -142,21 +159,15 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
     _geometry_column_name = None
 
     def __init__(
-        self,
-        data=None,
-        *args,
-        geometry: Any = None,
-        crs: Any | None = None,
-        **kwargs,
-    ) -> None:
-        with compat.ignore_shapely2_warnings():
-            if (
-                kwargs.get("copy") is None
-                and isinstance(data, DataFrame)
-                and not isinstance(data, GeoDataFrame)
-            ):
-                kwargs.update(copy=True)
-            super().__init__(data, *args, **kwargs)
+        self, data=None, *args, geometry: Any = None, crs: Any | None = None, **kwargs
+    ):
+        if (
+            kwargs.get("copy") is None
+            and isinstance(data, DataFrame)
+            and not isinstance(data, GeoDataFrame)
+        ):
+            kwargs.update(copy=True)
+        super().__init__(data, *args, **kwargs)
 
         # set_geometry ensures the geometry data have the proper dtype,
         # but is not called if `geometry=None` ('geometry' column present
@@ -208,6 +219,11 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
                 and not geometry.crs == crs
             ):
                 raise ValueError(crs_mismatch_error)
+
+            if hasattr(geometry, "name") and geometry.name not in ("geometry", None):
+                # __init__ always creates geometry col named "geometry"
+                # rename as `set_geometry` respects the given series name
+                geometry = geometry.rename("geometry")
 
             self.set_geometry(geometry, inplace=True, crs=crs)
 
@@ -273,8 +289,7 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         drop: bool = ...,
         inplace: Literal[True] = ...,
         crs: Any | None = ...,
-    ) -> None:
-        ...
+    ) -> None: ...
 
     @typing.overload
     def set_geometry(
@@ -283,11 +298,14 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         drop: bool = ...,
         inplace: Literal[False] = ...,
         crs: Any | None = ...,
-    ) -> GeoDataFrame:
-        ...
+    ) -> GeoDataFrame: ...
 
     def set_geometry(
-        self, col, drop: bool = False, inplace: bool = False, crs: Any | None = None
+        self,
+        col,
+        drop: bool | None = None,
+        inplace: bool = False,
+        crs: Any | None = None,
     ) -> GeoDataFrame | None:
         """
         Set the GeoDataFrame geometry using either an existing column or
@@ -297,9 +315,20 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
 
         Parameters
         ----------
-        col : column label or array
+        col : column label or array-like
+            An existing column name or values to set as the new geometry column.
+            If values (array-like, (Geo)Series) are passed, then if they are named
+            (Series) the new geometry column will have the corresponding name,
+            otherwise the existing geometry column will be replaced. If there is
+            no existing geometry column, the new geometry column will use the
+            default name "geometry".
         drop : boolean, default False
-            Delete column to be used as the new geometry
+            When specifying a named Series or an existing column name for `col`,
+            controls if the previous geometry column should be dropped from the
+            result. The default of False keeps both the old and new geometry column.
+
+            .. deprecated:: 1.0.0
+
         inplace : boolean, default False
             Modify the GeoDataFrame in place (do not create a new object)
         crs : pyproj.CRS, optional
@@ -315,25 +344,25 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
         >>> gdf = geopandas.GeoDataFrame(d, crs="EPSG:4326")
         >>> gdf
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        1  name2  POINT (2.00000 1.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
 
         Passing an array:
 
         >>> df1 = gdf.set_geometry([Point(0,0), Point(1,1)])
         >>> df1
-            col1                 geometry
-        0  name1  POINT (0.00000 0.00000)
-        1  name2  POINT (1.00000 1.00000)
+            col1     geometry
+        0  name1  POINT (0 0)
+        1  name2  POINT (1 1)
 
         Using existing column:
 
         >>> gdf["buffered"] = gdf.buffer(2)
         >>> df2 = gdf.set_geometry("buffered")
         >>> df2.geometry
-        0    POLYGON ((3.00000 2.00000, 2.99037 1.80397, 2....
-        1    POLYGON ((4.00000 1.00000, 3.99037 0.80397, 3....
+        0    POLYGON ((3 2, 2.99037 1.80397, 2.96157 1.6098...
+        1    POLYGON ((4 1, 3.99037 0.80397, 3.96157 0.6098...
         Name: buffered, dtype: geometry
 
         Returns
@@ -348,47 +377,79 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         if inplace:
             frame = self
         else:
-            frame = self.copy()
+            if PANDAS_GE_30:
+                frame = self.copy(deep=False)
+            else:
+                frame = self.copy()
 
-        to_remove = None
         geo_column_name = self._geometry_column_name
+
         if geo_column_name is None:
             geo_column_name = "geometry"
         if isinstance(col, (Series, list, np.ndarray, GeometryArray)):
+            if drop:
+                msg = (
+                    "The `drop` keyword argument is deprecated and has no effect when "
+                    "`col` is an array-like value. You should stop passing `drop` to "
+                    "`set_geometry` when this is the case."
+                )
+                warnings.warn(msg, category=FutureWarning, stacklevel=2)
+            if isinstance(col, Series) and col.name is not None:
+                geo_column_name = col.name
+
             level = col
         elif hasattr(col, "ndim") and col.ndim > 1:
             raise ValueError("Must pass array with one dimension only.")
-        else:
+        else:  # should be a colname
             try:
                 level = frame[col]
             except KeyError:
                 raise ValueError("Unknown column %s" % col)
-            except Exception:
-                raise
             if isinstance(level, DataFrame):
                 raise ValueError(
                     "GeoDataFrame does not support setting the geometry column where "
                     "the column name is shared by multiple columns."
                 )
 
-            if drop:
-                to_remove = col
-            else:
-                geo_column_name = col
+            given_colname_drop_msg = (
+                "The `drop` keyword argument is deprecated and in future the only "
+                "supported behaviour will match drop=False. To silence this "
+                "warning and adopt the future behaviour, stop providing "
+                "`drop` as a keyword to `set_geometry`. To replicate the "
+                "`drop=True` behaviour you should update "
+                "your code to\n`geo_col_name = gdf.active_geometry_name;"
+                " gdf.set_geometry(new_geo_col).drop("
+                "columns=geo_col_name).rename_geometry(geo_col_name)`."
+            )
 
-        if to_remove:
-            del frame[to_remove]
+            if drop is False:  # specifically False, not falsy i.e. None
+                # User supplied False explicitly, but arg is deprecated
+                warnings.warn(
+                    given_colname_drop_msg,
+                    category=FutureWarning,
+                    stacklevel=2,
+                )
+            if drop:
+                del frame[col]
+                warnings.warn(
+                    given_colname_drop_msg,
+                    category=FutureWarning,
+                    stacklevel=2,
+                )
+            else:
+                # if not dropping, set the active geometry name to the given col name
+                geo_column_name = col
 
         if not crs:
             crs = getattr(level, "crs", None)
 
-        if isinstance(level, (GeoSeries, GeometryArray)) and level.crs != crs:
-            # Avoids caching issues/crs sharing issues
-            level = level.copy()
-            level.crs = crs
-
         # Check that we are using a listlike of geometries
         level = _ensure_geometry(level, crs=crs)
+        # ensure_geometry only sets crs on level if it has crs==None
+        if isinstance(level, GeoSeries):
+            level.array.crs = crs
+        else:
+            level.crs = crs
         # update _geometry_column_name prior to assignment
         # to avoid default is None warning
         frame._geometry_column_name = geo_column_name
@@ -402,16 +463,14 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         self,
         col,
         inplace: Literal[True] = ...,
-    ) -> None:
-        ...
+    ) -> None: ...
 
     @typing.overload
     def rename_geometry(
         self,
         col,
         inplace: Literal[False] = ...,
-    ) -> GeoDataFrame:
-        ...
+    ) -> GeoDataFrame: ...
 
     def rename_geometry(self, col, inplace: bool = False) -> GeoDataFrame | None:
         """
@@ -449,10 +508,30 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         else:
             if not inplace:
                 return self.rename(columns={geometry_col: col}).set_geometry(
-                    col, inplace
+                    col, inplace=inplace
                 )
             self.rename(columns={geometry_col: col}, inplace=inplace)
             self.set_geometry(col, inplace=inplace)
+
+    @property
+    def active_geometry_name(self):
+        """Return the name of the active geometry column
+
+        Returns a string name if a GeoDataFrame has an active geometry column set.
+        Otherwise returns None. You can also access the active geometry column using the
+        ``.geometry`` property. You can set a GeoSeries to be an active geometry
+        using the :meth:`~GeoDataFrame.set_geometry` method.
+
+        Returns
+        -------
+        str
+            name of an active geometry column or None
+
+        See also
+        --------
+        GeoDataFrame.set_geometry : set the active geometry
+        """
+        return self._geometry_column_name
 
     @property
     def crs(self) -> CRS:
@@ -508,6 +587,14 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
             )
 
         if hasattr(self.geometry.values, "crs"):
+            if self.crs is not None:
+                warnings.warn(
+                    "Overriding the CRS of a GeoDataFrame that already has CRS. "
+                    "This unsafe behavior will be deprecated in future versions. "
+                    "Use GeoDataFrame.set_crs method instead",
+                    stacklevel=2,
+                    category=DeprecationWarning,
+                )
             self.geometry.values.crs = value
         else:
             # column called 'geometry' without geometry
@@ -525,7 +612,15 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
                 crs = state.pop("crs", None)
             else:
                 crs = state.pop("_crs", None)
-            crs = CRS.from_user_input(crs) if crs is not None else crs
+            if crs is not None and not HAS_PYPROJ:
+                raise ImportError(
+                    "Unpickling a GeoDataFrame with CRS requires the 'pyproj' package, "
+                    "but it is not installed or does not import correctly. "
+                )
+            elif crs is not None:
+                from pyproj import CRS
+
+                crs = CRS.from_user_input(crs)
 
         super().__setstate__(state)
 
@@ -578,17 +673,17 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         It is recommended to use :func:`geopandas.read_file` instead.
 
         Can load a ``GeoDataFrame`` from a file in any format recognized by
-        `fiona`. See http://fiona.readthedocs.io/en/latest/manual.html for details.
+        `pyogrio`. See http://pyogrio.readthedocs.io/ for details.
 
         Parameters
         ----------
         filename : str
             File path or file handle to read from. Depending on which kwargs
             are included, the content of filename may vary. See
-            http://fiona.readthedocs.io/en/latest/README.html#usage for usage details.
+            :func:`pyogrio.read_dataframe` for usage details.
         kwargs : key-word arguments
-            These arguments are passed to fiona.open, and can be used to
-            access multi-layer data, data stored within archives (zip files),
+            These arguments are passed to :func:`pyogrio.read_dataframe`, and can be
+            used to access multi-layer data, data stored within archives (zip files),
             etc.
 
         Examples
@@ -679,9 +774,9 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         ... }
         >>> df = geopandas.GeoDataFrame.from_features(feature_coll)
         >>> df
-                          geometry   col1
-        0  POINT (1.00000 2.00000)  name1
-        1  POINT (2.00000 1.00000)  name2
+              geometry   col1
+        0  POINT (1 2)  name1
+        1  POINT (2 1)  name2
 
         """
         # Handle feature collections
@@ -795,6 +890,42 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
 
         return df
 
+    @classmethod
+    def from_arrow(cls, table, geometry=None):
+        """
+        Construct a GeoDataFrame from a Arrow table object based on GeoArrow
+        extension types.
+
+        See https://geoarrow.org/ for details on the GeoArrow specification.
+
+        This functions accepts any tabular Arrow object implementing
+        the `Arrow PyCapsule Protocol`_ (i.e. having an ``__arrow_c_array__``
+        or ``__arrow_c_stream__`` method).
+
+        .. _Arrow PyCapsule Protocol: https://arrow.apache.org/docs/format/CDataInterface/PyCapsuleInterface.html
+
+        .. versionadded:: 1.0
+
+        Parameters
+        ----------
+        table : pyarrow.Table or Arrow-compatible table
+            Any tabular object implementing the Arrow PyCapsule Protocol
+            (i.e. has an ``__arrow_c_array__`` or ``__arrow_c_stream__``
+            method). This table should have at least one column with a
+            geoarrow geometry type.
+        geometry : str, default None
+            The name of the geometry column to set as the active geometry
+            column. If None, the first geometry column found will be used.
+
+        Returns
+        -------
+        GeoDataFrame
+
+        """
+        from geopandas.io._geoarrow import arrow_to_geopandas
+
+        return arrow_to_geopandas(table, geometry=geometry)
+
     def to_json(
         self,
         na: Literal["null", "drop", "keep"] = "null",
@@ -847,9 +978,9 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
         >>> gdf = geopandas.GeoDataFrame(d, crs="EPSG:3857")
         >>> gdf
-            col1             geometry
-        0  name1  POINT (1.000 2.000)
-        1  name2  POINT (2.000 1.000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
 
         >>> gdf.to_json()
         '{"type": "FeatureCollection", "features": [{"id": "0", "type": "Feature", \
@@ -877,7 +1008,7 @@ es": {"name": "urn:ogc:def:crs:EPSG::3857"}}}'
         else:
             df = self
 
-        geo = df._to_geo(na=na, show_bbox=show_bbox, drop_id=drop_id)
+        geo = df.to_geo_dict(na=na, show_bbox=show_bbox, drop_id=drop_id)
 
         # if the geometry is not in WGS84, include CRS in the JSON
         if df.crs is not None and not df.crs.equals("epsg:4326"):
@@ -905,7 +1036,7 @@ es": {"name": "urn:ogc:def:crs:EPSG::3857"}}}'
         represents the ``GeoDataFrame`` as a GeoJSON-like
         ``FeatureCollection``.
 
-        This differs from `_to_geo()` only in that it is a property with
+        This differs from :meth:`to_geo_dict` only in that it is a property with
         default args instead of a method.
 
         CRS of the dataframe is not passed on to the output, unlike
@@ -918,9 +1049,9 @@ es": {"name": "urn:ogc:def:crs:EPSG::3857"}}}'
         >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
         >>> gdf = geopandas.GeoDataFrame(d, crs="EPSG:4326")
         >>> gdf
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        1  name2  POINT (2.00000 1.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
 
         >>> gdf.__geo_interface__
         {'type': 'FeatureCollection', 'features': [{'id': '0', 'type': 'Feature', \
@@ -929,7 +1060,7 @@ es": {"name": "urn:ogc:def:crs:EPSG::3857"}}}'
 ': {'col1': 'name2'}, 'geometry': {'type': 'Point', 'coordinates': (2.0, 1.0)}, 'b\
 box': (2.0, 1.0, 2.0, 1.0)}], 'bbox': (1.0, 1.0, 2.0, 2.0)}
         """
-        return self._to_geo(na="null", show_bbox=True, drop_id=False)
+        return self.to_geo_dict(na="null", show_bbox=True, drop_id=False)
 
     def iterfeatures(
         self, na: str = "null", show_bbox: bool = False, drop_id: bool = False
@@ -963,9 +1094,9 @@ individually so that features may have different properties
         >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
         >>> gdf = geopandas.GeoDataFrame(d, crs="EPSG:4326")
         >>> gdf
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        1  name2  POINT (2.00000 1.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
 
         >>> feature = next(gdf.iterfeatures())
         >>> feature
@@ -977,8 +1108,8 @@ individually so that features may have different properties
 
         if self._geometry_column_name not in self:
             raise AttributeError(
-                "No geometry data set (expected in"
-                " column '%s')." % self._geometry_column_name
+                "No geometry data set (expected in column '%s')."
+                % self._geometry_column_name
             )
 
         ids = np.array(self.index, copy=False)
@@ -992,13 +1123,13 @@ individually so that features may have different properties
         if len(properties_cols) > 0:
             # convert to object to get python scalars.
             properties_cols = self[properties_cols]
-            properties = properties_cols.astype(object).values
+            properties = properties_cols.astype(object)
             na_mask = pd.isna(properties_cols).values
 
             if na == "null":
                 properties[na_mask] = None
 
-            for i, row in enumerate(properties):
+            for i, row in enumerate(properties.values):
                 geom = geometries[i]
 
                 if na == "drop":
@@ -1041,18 +1172,60 @@ individually so that features may have different properties
 
                 yield feature
 
-    def _to_geo(self, **kwargs) -> dict:
+    def to_geo_dict(self, na="null", show_bbox=False, drop_id=False) -> dict:
         """
-        Returns a python feature collection (i.e. the geointerface)
-        representation of the GeoDataFrame.
+        Returns a python feature collection representation of the GeoDataFrame
+        as a dictionary with a list of features based on the ``__geo_interface__``
+        GeoJSON-like specification.
+
+        Parameters
+        ----------
+        na : str, optional
+            Options are {'null', 'drop', 'keep'}, default 'null'.
+            Indicates how to output missing (NaN) values in the GeoDataFrame
+
+            - null: output the missing entries as JSON null
+            - drop: remove the property from the feature. This applies to each feature \
+individually so that features may have different properties
+            - keep: output the missing entries as NaN
+
+        show_bbox : bool, optional
+            Include bbox (bounds) in the geojson. Default False.
+        drop_id : bool, default: False
+            Whether to retain the index of the GeoDataFrame as the id property
+            in the generated dictionary. Default is False, but may want True
+            if the index is just arbitrary row numbers.
+
+        Examples
+        --------
+
+        >>> from shapely.geometry import Point
+        >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
+        >>> gdf = geopandas.GeoDataFrame(d)
+        >>> gdf
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
+
+        >>> gdf.to_geo_dict()
+        {'type': 'FeatureCollection', 'features': [{'id': '0', 'type': 'Feature', '\
+properties': {'col1': 'name1'}, 'geometry': {'type': 'Point', 'coordinates': (1.0, \
+2.0)}}, {'id': '1', 'type': 'Feature', 'properties': {'col1': 'name2'}, 'geometry':\
+ {'type': 'Point', 'coordinates': (2.0, 1.0)}}]}
+
+        See also
+        --------
+        GeoDataFrame.to_json : return a GeoDataFrame as a GeoJSON string
 
         """
         geo = {
             "type": "FeatureCollection",
-            "features": list(self.iterfeatures(**kwargs)),
+            "features": list(
+                self.iterfeatures(na=na, show_bbox=show_bbox, drop_id=drop_id)
+            ),
         }
 
-        if kwargs.get("show_bbox", False):
+        if show_bbox:
             geo["bbox"] = tuple(self.total_bounds)
 
         return geo
@@ -1068,8 +1241,7 @@ individually so that features may have different properties
             The default is to return a binary bytes object.
         kwargs
             Additional keyword args will be passed to
-            :func:`shapely.to_wkb` if shapely >= 2 is installed or
-            :func:`pygeos.to_wkb` if pygeos is installed.
+            :func:`shapely.to_wkb`.
 
         Returns
         -------
@@ -1092,8 +1264,7 @@ individually so that features may have different properties
         Parameters
         ----------
         kwargs
-            Keyword args will be passed to :func:`pygeos.to_wkt`
-            if pygeos is installed.
+            Keyword args will be passed to :func:`shapely.to_wkt`.
 
         Returns
         -------
@@ -1109,28 +1280,110 @@ individually so that features may have different properties
 
         return df
 
+    def to_arrow(
+        self, *, index=None, geometry_encoding="WKB", interleaved=True, include_z=None
+    ):
+        """Encode a GeoDataFrame to GeoArrow format.
+
+        See https://geoarrow.org/ for details on the GeoArrow specification.
+
+        This functions returns a generic Arrow data object implementing
+        the `Arrow PyCapsule Protocol`_ (i.e. having an ``__arrow_c_stream__``
+        method). This object can then be consumed by your Arrow implementation
+        of choice that supports this protocol.
+
+        .. _Arrow PyCapsule Protocol: https://arrow.apache.org/docs/format/CDataInterface/PyCapsuleInterface.html
+
+        .. versionadded:: 1.0
+
+        Parameters
+        ----------
+        index : bool, default None
+            If ``True``, always include the dataframe's index(es) as columns
+            in the file output.
+            If ``False``, the index(es) will not be written to the file.
+            If ``None``, the index(ex) will be included as columns in the file
+            output except `RangeIndex` which is stored as metadata only.
+        geometry_encoding : {'WKB', 'geoarrow' }, default 'WKB'
+            The GeoArrow encoding to use for the data conversion.
+        interleaved : bool, default True
+            Only relevant for 'geoarrow' encoding. If True, the geometries'
+            coordinates are interleaved in a single fixed size list array.
+            If False, the coordinates are stored as separate arrays in a
+            struct type.
+        include_z : bool, default None
+            Only relevant for 'geoarrow' encoding (for WKB, the dimensionality
+            of the individial geometries is preserved).
+            If False, return 2D geometries. If True, include the third dimension
+            in the output (if a geometry has no third dimension, the z-coordinates
+            will be NaN). By default, will infer the dimensionality from the
+            input geometries. Note that this inference can be unreliable with
+            empty geometries (for a guaranteed result, it is recommended to
+            specify the keyword).
+
+        Returns
+        -------
+        ArrowTable
+            A generic Arrow table object with geometry columns encoded to
+            GeoArrow.
+
+        Examples
+        --------
+        >>> from shapely.geometry import Point
+        >>> data = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
+        >>> gdf = geopandas.GeoDataFrame(data)
+        >>> gdf
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
+
+        >>> arrow_table = gdf.to_arrow()
+        >>> arrow_table
+        <geopandas.io._geoarrow.ArrowTable object at ...>
+
+        The returned data object needs to be consumed by a library implementing
+        the Arrow PyCapsule Protocol. For example, wrapping the data as a
+        pyarrow.Table (requires pyarrow >= 14.0):
+
+        >>> import pyarrow as pa
+        >>> table = pa.table(arrow_table)
+        >>> table
+        pyarrow.Table
+        col1: string
+        geometry: binary
+        ----
+        col1: [["name1","name2"]]
+        geometry: [[0101000000000000000000F03F0000000000000040,\
+01010000000000000000000040000000000000F03F]]
+
+        """
+        from geopandas.io._geoarrow import ArrowTable, geopandas_to_arrow
+
+        table, _ = geopandas_to_arrow(
+            self,
+            index=index,
+            geometry_encoding=geometry_encoding,
+            interleaved=interleaved,
+            include_z=include_z,
+        )
+        return ArrowTable(table)
+
     def to_parquet(
         self,
         path: os.PathLike | typing.IO,
         index: bool | None = None,
         compression: str = "snappy",
+        geometry_encoding="WKB",
+        write_covering_bbox=False,
         schema_version: SUPPORTED_VERSIONS_LITERAL | None = None,
         **kwargs,
     ) -> None:
         """Write a GeoDataFrame to the Parquet format.
 
-        Any geometry columns present are serialized to WKB format in the file.
+        By default, all geometry columns present are serialized to WKB format
+        in the file.
 
         Requires 'pyarrow'.
-
-        WARNING: this is an early implementation of Parquet file support and
-        associated metadata, the specification for which continues to evolve.
-        This is tracking version 0.4.0 of the GeoParquet specification at:
-        https://github.com/opengeospatial/geoparquet
-
-        This metadata specification does not yet make stability promises.  As such,
-        we do not yet recommend using this in a production setting unless you are
-        able to rewrite your Parquet files.
 
         .. versionadded:: 0.8
 
@@ -1145,9 +1398,25 @@ individually so that features may have different properties
             output except `RangeIndex` which is stored as metadata only.
         compression : {'snappy', 'gzip', 'brotli', None}, default 'snappy'
             Name of the compression to use. Use ``None`` for no compression.
-        schema_version : {'0.1.0', '0.4.0', None}
-            GeoParquet specification version; if not provided will default to
-            latest supported version.
+        geometry_encoding : {'WKB', 'geoarrow'}, default 'WKB'
+            The encoding to use for the geometry columns. Defaults to "WKB"
+            for maximum interoperability. Specify "geoarrow" to use one of the
+            native GeoArrow-based single-geometry type encodings.
+            Note: the "geoarrow" option is part of the newer GeoParquet 1.1
+            specification, should be considered as experimental, and may not
+            be supported by all readers.
+        write_covering_bbox : bool, default False
+            Writes the bounding box column for each row entry with column
+            name 'bbox'. Writing a bbox column can be computationally
+            expensive, but allows you to specify a `bbox` in :
+            func:`read_parquet` for filtered reading.
+            Note: this bbox column is part of the newer GeoParquet 1.1
+            specification and should be considered as experimental. While
+            writing the column is backwards compatible, using it for filtering
+            may not be supported by all readers.
+        schema_version : {'0.1.0', '0.4.0', '1.0.0', '1.1.0', None}
+            GeoParquet specification version; if not provided, will default to
+            latest supported stable version (1.0.0).
         kwargs
             Additional keyword arguments passed to to pyarrow.parquet.write_table().
 
@@ -1168,7 +1437,7 @@ individually so that features may have different properties
         engine = kwargs.pop("engine", "auto")
         if engine not in ("auto", "pyarrow"):
             raise ValueError(
-                f"GeoPandas only supports using pyarrow as the engine for "
+                "GeoPandas only supports using pyarrow as the engine for "
                 f"to_parquet: {engine!r} passed instead."
             )
 
@@ -1178,8 +1447,10 @@ individually so that features may have different properties
             self,
             path,
             compression=compression,
+            geometry_encoding=geometry_encoding,
             index=index,
             schema_version=schema_version,
+            write_covering_bbox=write_covering_bbox,
             **kwargs,
         )
 
@@ -1197,15 +1468,6 @@ individually so that features may have different properties
 
         Requires 'pyarrow' >= 0.17.
 
-        WARNING: this is an early implementation of Parquet file support and
-        associated metadata, the specification for which continues to evolve.
-        This is tracking version 0.4.0 of the GeoParquet specification at:
-        https://github.com/opengeospatial/geoparquet
-
-        This metadata specification does not yet make stability promises.  As such,
-        we do not yet recommend using this in a production setting unless you are
-        able to rewrite your Feather files.
-
         .. versionadded:: 0.8
 
         Parameters
@@ -1220,7 +1482,7 @@ individually so that features may have different properties
         compression : {'zstd', 'lz4', 'uncompressed'}, optional
             Name of the compression to use. Use ``"uncompressed"`` for no
             compression. By default uses LZ4 if available, otherwise uncompressed.
-        schema_version : {'0.1.0', '0.4.0', None}
+        schema_version : {'0.1.0', '0.4.0', '1.0.0', None}
             GeoParquet specification version; if not provided will default to
             latest supported version.
         kwargs
@@ -1259,11 +1521,11 @@ individually so that features may have different properties
         """Write the ``GeoDataFrame`` to a file.
 
         By default, an ESRI shapefile is written, but any OGR data source
-        supported by Fiona can be written. A dictionary of supported OGR
+        supported by Pyogrio or Fiona can be written. A dictionary of supported OGR
         providers is available via:
 
-        >>> import fiona
-        >>> fiona.supported_drivers  # doctest: +SKIP
+        >>> import pyogrio
+        >>> pyogrio.list_drivers()  # doctest: +SKIP
 
         Parameters
         ----------
@@ -1298,18 +1560,22 @@ individually so that features may have different properties
             will determine the crs based on crs df attribute.
             The value can be anything accepted
             by :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
-            such as an authority string (eg "EPSG:4326") or a WKT string.
-        engine : str, "fiona" or "pyogrio"
+            such as an authority string (eg "EPSG:4326") or a WKT string. The keyword
+            is not supported for the "pyogrio" engine.
+        engine : str, "pyogrio" or "fiona"
             The underlying library that is used to write the file. Currently, the
-            supported options are "fiona" and "pyogrio". Defaults to "fiona" if
-            installed, otherwise tries "pyogrio".
+            supported options are "pyogrio" and "fiona". Defaults to "pyogrio" if
+            installed, otherwise tries "fiona".
+        metadata : dict[str, str], default None
+            Optional metadata to be stored in the file. Keys and values must be
+            strings. Supported only for "GPKG" driver.
         **kwargs :
             Keyword args to be passed to the engine, and can be used to write
             to multi-layer data, store data within archives (zip files), etc.
-            In case of the "fiona" engine, the keyword arguments are passed to
-            fiona.open`. For more information on possible keywords, type:
-            ``import fiona; help(fiona.open)``. In case of the "pyogrio" engine,
-            the keyword arguments are passed to `pyogrio.write_dataframe`.
+            In case of the "pyogrio" engine, the keyword arguments are passed to
+            `pyogrio.write_dataframe`. In case of the "fiona" engine, the keyword
+            arguments are passed to fiona.open`. For more information on possible
+            keywords, type: ``import pyogrio; help(pyogrio.write_dataframe)``.
 
         Notes
         -----
@@ -1356,8 +1622,7 @@ individually so that features may have different properties
         epsg: int | None = ...,
         inplace: Literal[True] = ...,
         allow_override: bool = ...,
-    ) -> None:
-        ...
+    ) -> None: ...
 
     @typing.overload
     def set_crs(
@@ -1366,8 +1631,7 @@ individually so that features may have different properties
         epsg: int | None = ...,
         inplace: Literal[False] = ...,
         allow_override: bool = ...,
-    ) -> GeoDataFrame:
-        ...
+    ) -> GeoDataFrame: ...
 
     def set_crs(
         self,
@@ -1382,16 +1646,20 @@ individually so that features may have different properties
         If there are multiple geometry columns within the GeoDataFrame, only
         the CRS of the active geometry column is set.
 
-        NOTE: The underlying geometries are not transformed to this CRS. To
+        Pass ``None`` to remove CRS from the active geometry column.
+
+        Notes
+        -----
+        The underlying geometries are not transformed to this CRS. To
         transform the geometries to a new CRS, use the ``to_crs`` method.
 
         Parameters
         ----------
-        crs : pyproj.CRS, optional if `epsg` is specified
+        crs : pyproj.CRS | None, optional
             The value can be anything accepted
             by :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
             such as an authority string (eg "EPSG:4326") or a WKT string.
-        epsg : int, optional if `crs` is specified
+        epsg : int, optional
             EPSG code specifying the projection.
         inplace : bool, default False
             If True, the CRS of the GeoDataFrame will be changed in place
@@ -1407,9 +1675,9 @@ individually so that features may have different properties
         >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
         >>> gdf = geopandas.GeoDataFrame(d)
         >>> gdf
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        1  name2  POINT (2.00000 1.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
 
         Setting CRS to a GeoDataFrame without one:
 
@@ -1460,8 +1728,7 @@ individually so that features may have different properties
         crs: Any | None = ...,
         epsg: int | None = ...,
         inplace: Literal[False] = ...,
-    ) -> GeoDataFrame:
-        ...
+    ) -> GeoDataFrame: ...
 
     @typing.overload
     def to_crs(
@@ -1469,8 +1736,7 @@ individually so that features may have different properties
         crs: Any | None = ...,
         epsg: int | None = ...,
         inplace: Literal[True] = ...,
-    ) -> None:
-        ...
+    ) -> None: ...
 
     def to_crs(
         self,
@@ -1512,9 +1778,9 @@ individually so that features may have different properties
         >>> d = {'col1': ['name1', 'name2'], 'geometry': [Point(1, 2), Point(2, 1)]}
         >>> gdf = geopandas.GeoDataFrame(d, crs=4326)
         >>> gdf
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        1  name2  POINT (2.00000 1.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
         >>> gdf.crs  # doctest: +SKIP
         <Geographic 2D CRS: EPSG:4326>
         Name: WGS 84
@@ -1566,8 +1832,6 @@ individually so that features may have different properties
         """Returns the estimated UTM CRS based on the bounds of the dataset.
 
         .. versionadded:: 0.9
-
-        .. note:: Requires pyproj 3+
 
         Parameters
         ----------
@@ -1665,7 +1929,10 @@ individually so that features may have different properties
             if pd.api.types.is_scalar(value) or isinstance(value, BaseGeometry):
                 value = [value] * self.shape[0]
             try:
-                crs = getattr(self, "crs", None)
+                if self._geometry_column_name is not None:
+                    crs = getattr(self, "crs", None)
+                else:  # don't use getattr, because a col "crs" might exist
+                    crs = None
                 value = _ensure_geometry(value, crs=crs)
                 if key == "geometry":
                     self._persist_old_default_geometry_colname()
@@ -1687,34 +1954,6 @@ individually so that features may have different properties
             copied._geometry_column_name = self._geometry_column_name
         return copied
 
-    def merge(self, *args, **kwargs) -> pd.DataFrame | GeoDataFrame:
-        r"""Merge two ``GeoDataFrame`` objects with a database-style join.
-
-        Returns a ``GeoDataFrame`` if a geometry column is present; otherwise,
-        returns a pandas ``DataFrame``.
-
-        Returns
-        -------
-        GeoDataFrame or DataFrame
-
-        Notes
-        -----
-        The extra arguments ``*args`` and keyword arguments ``**kwargs`` are
-        passed to DataFrame.merge.
-        See https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas\
-.DataFrame.merge.html
-        for more details.
-        """
-        result = DataFrame.merge(self, *args, **kwargs)
-        geo_col = self._geometry_column_name
-        if isinstance(result, DataFrame) and geo_col in result:
-            result.__class__ = GeoDataFrame
-            result.crs = self.crs
-            result._geometry_column_name = geo_col
-        elif isinstance(result, DataFrame) and geo_col not in result:
-            result.__class__ = DataFrame
-        return result
-
     @doc(pd.DataFrame)
     def apply(
         self,
@@ -1728,13 +1967,6 @@ individually so that features may have different properties
         result = super().apply(
             func, axis=axis, raw=raw, result_type=result_type, args=args, **kwargs
         )
-        # pandas <1.4 re-attach last geometry col if lost
-        if (
-            not compat.PANDAS_GE_14
-            and isinstance(result, GeoDataFrame)
-            and result._geometry_column_name is None
-        ):
-            result._geometry_column_name = self._geometry_column_name
         # Reconstruct gdf if it was lost by apply
         if (
             isinstance(result, DataFrame)
@@ -1757,7 +1989,7 @@ individually so that features may have different properties
                     # not enough info about func to preserve CRS
                     result = _ensure_geometry(result)
 
-                except TypeError:
+                except (TypeError, shapely.errors.GeometryTypeError):
                     pass
 
         return result
@@ -1765,6 +1997,20 @@ individually so that features may have different properties
     @property
     def _constructor(self) -> DataFrame | GeoDataFrame:
         return _geodataframe_constructor_with_fallback
+
+    def _constructor_from_mgr(self, mgr, axes):
+        # replicate _geodataframe_constructor_with_fallback behaviour
+        # unless safe to skip
+        if not any(isinstance(block.dtype, GeometryDtype) for block in mgr.blocks):
+            return _geodataframe_constructor_with_fallback(
+                pd.DataFrame._from_mgr(mgr, axes)
+            )
+        gdf = GeoDataFrame._from_mgr(mgr, axes)
+        # _from_mgr doesn't preserve metadata (expect __finalize__ to be called)
+        # still need to mimic __init__ behaviour with geometry=None
+        if (gdf.columns == "geometry").sum() == 1:  # only if "geometry" is single col
+            gdf._geometry_column_name = "geometry"
+        return gdf
 
     @property
     def _constructor_sliced(self) -> Series | GeoSeries:
@@ -1785,14 +2031,21 @@ individually so that features may have different properties
               checking the identity of the index)
             """
             srs = pd.Series(*args, **kwargs)
-            is_row_proxy = srs.index is self.columns
+            is_row_proxy = srs.index.is_(self.columns)
             if is_geometry_type(srs) and not is_row_proxy:
                 srs = GeoSeries(srs)
             return srs
 
         return _geodataframe_constructor_sliced
 
-    def __finalize__(self, other, method: str | None = None, **kwargs) -> GeoDataFrame:
+    def _constructor_sliced_from_mgr(self, mgr, axes):
+        is_row_proxy = mgr.index.is_(self.columns)
+
+        if isinstance(mgr.blocks[0].dtype, GeometryDtype) and not is_row_proxy:
+            return GeoSeries._from_mgr(mgr, axes)
+        return Series._from_mgr(mgr, axes)
+
+    def __finalize__(self, other, method: str | None = None, **kwargs):
         """propagate metadata from other to self"""
         self = super().__finalize__(other, method=method, **kwargs)
 
@@ -1809,8 +2062,8 @@ individually so that features may have different properties
                 raise ValueError(
                     "Concat operation has resulted in multiple columns using "
                     f"the geometry column name '{self._geometry_column_name}'.\n"
-                    f"Please ensure this column from the first DataFrame is not "
-                    f"repeated."
+                    "Please ensure this column from the first DataFrame is not "
+                    "repeated."
                 )
         elif method == "unstack":
             # unstack adds multiindex columns and reshapes data.
@@ -1828,11 +2081,12 @@ individually so that features may have different properties
         sort: bool = True,
         observed: bool = False,
         dropna: bool = True,
+        method="unary",
         **kwargs,
     ) -> GeoDataFrame:
         """
         Dissolve geometries within `groupby` into single observation.
-        This is accomplished by applying the `unary_union` method
+        This is accomplished by applying the `union_all` method
         to all geometries within a groupself.
 
         Observations associated with each `groupby` group will be aggregated
@@ -1840,9 +2094,12 @@ individually so that features may have different properties
 
         Parameters
         ----------
-        by : string, default None
-            Column whose values define groups to be dissolved. If None,
-            whole GeoDataFrame is considered a single group.
+        by : str or list-like, default None
+            Column(s) whose values define the groups to be dissolved. If None,
+            the entire GeoDataFrame is considered as a single group. If a list-like
+            object is provided, the values in the list are treated as categorical
+            labels, and polygons will be combined based on the equality of
+            these categorical labels.
         aggfunc : function or string, default "first"
             Aggregation function for manipulation of data associated
             with each group. Passed to pandas `groupby.agg` method.
@@ -1857,26 +2114,28 @@ individually so that features may have different properties
         level : int or str or sequence of int or sequence of str, default None
             If the axis is a MultiIndex (hierarchical), group by a
             particular level or levels.
-
-            .. versionadded:: 0.9.0
         sort : bool, default True
             Sort group keys. Get better performance by turning this off.
             Note this does not influence the order of observations within
             each group. Groupby preserves the order of rows within each group.
-
-            .. versionadded:: 0.9.0
         observed : bool, default False
             This only applies if any of the groupers are Categoricals.
             If True: only show observed values for categorical groupers.
             If False: show all values for categorical groupers.
-
-            .. versionadded:: 0.9.0
         dropna : bool, default True
             If True, and if group keys contain NA values, NA values
             together with row/column will be dropped. If False, NA
             values will also be treated as the key in groups.
+        method : str (default ``"unary"``)
+            The method to use for the union. Options are:
 
-            .. versionadded:: 0.9.0
+            * ``"unary"``: use the unary union algorithm. This option is the most robust
+              but can be slow for large numbers of geometries (default).
+            * ``"coverage"``: use the coverage union algorithm. This option is optimized
+              for non-overlapping polygons and can be significantly faster than the
+              unary union algorithm. However, it can produce invalid geometries if the
+              polygons overlap.
+
         **kwargs :
             Keyword arguments to be passed to the pandas `DataFrameGroupby.agg` method
             which is used by `dissolve`. In particular, `numeric_only` may be
@@ -1896,17 +2155,17 @@ individually so that features may have different properties
         ... }
         >>> gdf = geopandas.GeoDataFrame(d, crs=4326)
         >>> gdf
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        1  name2  POINT (2.00000 1.00000)
-        2  name1  POINT (0.00000 1.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name2  POINT (2 1)
+        2  name1  POINT (0 1)
 
         >>> dissolved = gdf.dissolve('col1')
         >>> dissolved  # doctest: +SKIP
-                                                    geometry
+                                geometry
         col1
-        name1  MULTIPOINT (0.00000 1.00000, 1.00000 2.00000)
-        name2                        POINT (2.00000 1.00000)
+        name1  MULTIPOINT ((0 1), (1 2))
+        name2                POINT (2 1)
 
         See also
         --------
@@ -1951,7 +2210,7 @@ individually so that features may have different properties
 
         # Process spatial component
         def merge_geometries(block):
-            merged_geom = block.unary_union
+            merged_geom = block.union_all(method=method)
             return merged_geom
 
         g = self.groupby(group_keys=False, **groupby_kwargs)[self.geometry.name].agg(
@@ -1974,7 +2233,7 @@ individually so that features may have different properties
         self,
         column=None,
         ignore_index: bool = False,
-        index_parts: bool | None = None,
+        index_parts: bool = False,
         **kwargs,
     ) -> GeoDataFrame | DataFrame:
         """
@@ -1993,7 +2252,7 @@ individually so that features may have different properties
         ignore_index : bool, default False
             If True, the resulting index will be labelled 0, 1, …, n - 1,
             ignoring `index_parts`.
-        index_parts : boolean, default True
+        index_parts : boolean, default False
             If True, the resulting index will be a multi-index (original
             index with an additional level indicating the multiple
             geometries: a new zero-based index for each single part geometry
@@ -2018,33 +2277,33 @@ individually so that features may have different properties
         ... }
         >>> gdf = geopandas.GeoDataFrame(d, crs=4326)
         >>> gdf
-            col1                                       geometry
-        0  name1  MULTIPOINT (1.00000 2.00000, 3.00000 4.00000)
-        1  name2  MULTIPOINT (2.00000 1.00000, 0.00000 0.00000)
+            col1               geometry
+        0  name1  MULTIPOINT ((1 2), (3 4))
+        1  name2  MULTIPOINT ((2 1), (0 0))
 
         >>> exploded = gdf.explode(index_parts=True)
         >>> exploded
-              col1                 geometry
-        0 0  name1  POINT (1.00000 2.00000)
-          1  name1  POINT (3.00000 4.00000)
-        1 0  name2  POINT (2.00000 1.00000)
-          1  name2  POINT (0.00000 0.00000)
+              col1     geometry
+        0 0  name1  POINT (1 2)
+          1  name1  POINT (3 4)
+        1 0  name2  POINT (2 1)
+          1  name2  POINT (0 0)
 
         >>> exploded = gdf.explode(index_parts=False)
         >>> exploded
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        0  name1  POINT (3.00000 4.00000)
-        1  name2  POINT (2.00000 1.00000)
-        1  name2  POINT (0.00000 0.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        0  name1  POINT (3 4)
+        1  name2  POINT (2 1)
+        1  name2  POINT (0 0)
 
         >>> exploded = gdf.explode(ignore_index=True)
         >>> exploded
-            col1                 geometry
-        0  name1  POINT (1.00000 2.00000)
-        1  name1  POINT (3.00000 4.00000)
-        2  name2  POINT (2.00000 1.00000)
-        3  name2  POINT (0.00000 0.00000)
+            col1     geometry
+        0  name1  POINT (1 2)
+        1  name1  POINT (3 4)
+        2  name2  POINT (2 1)
+        3  name2  POINT (0 0)
 
         See also
         --------
@@ -2058,18 +2317,6 @@ individually so that features may have different properties
         # If the specified column is not a geometry dtype use pandas explode
         if not isinstance(self[column].dtype, GeometryDtype):
             return super().explode(column, ignore_index=ignore_index, **kwargs)
-
-        if index_parts is None:
-            if not ignore_index:
-                warnings.warn(
-                    "Currently, index_parts defaults to True, but in the future, "
-                    "it will default to False to be consistent with Pandas. "
-                    "Use `index_parts=True` to keep the current behavior and "
-                    "True/False to silence the warning.",
-                    FutureWarning,
-                    stacklevel=2,
-                )
-            index_parts = True
 
         exploded_geom = self.geometry.reset_index(drop=True).explode(index_parts=True)
 
@@ -2094,22 +2341,25 @@ individually so that features may have different properties
         return df
 
     # overrides the pandas astype method to ensure the correct return type
+    # should be removable when pandas 1.4 is dropped
     def astype(
-        self, dtype, copy: bool = True, errors="raise", **kwargs
+        self, dtype, copy: bool | None = None, errors="raise", **kwargs
     ) -> pd.DataFrame | GeoDataFrame:
         """
         Cast a pandas object to a specified dtype ``dtype``.
-
         Returns a GeoDataFrame when the geometry column is kept as geometries,
         otherwise returns a pandas DataFrame.
-
         See the pandas.DataFrame.astype docstring for more details.
-
         Returns
         -------
         GeoDataFrame or DataFrame
         """
-        df = super().astype(dtype, copy=copy, errors=errors, **kwargs)
+        if not PANDAS_GE_30 and copy is None:
+            copy = True
+        if copy is not None:
+            kwargs["copy"] = copy
+
+        df = super().astype(dtype, errors=errors, **kwargs)
 
         try:
             geoms = df[self._geometry_column_name]
@@ -2120,29 +2370,6 @@ individually so that features may have different properties
         # if the geometry column is converted to non-geometries or did not exist
         # do not return a GeoDataFrame
         return pd.DataFrame(df)
-
-    def convert_dtypes(self, *args, **kwargs) -> GeoDataFrame:
-        """
-        Convert columns to best possible dtypes using dtypes supporting ``pd.NA``.
-
-        Always returns a GeoDataFrame as no conversions are applied to the
-        geometry column.
-
-        See the pandas.DataFrame.convert_dtypes docstring for more details.
-
-        Returns
-        -------
-        GeoDataFrame
-
-        """
-        # Overridden to fix GH1870, that return type is not preserved always
-        # (and where it was, geometry col was not)
-
-        return GeoDataFrame(
-            super().convert_dtypes(*args, **kwargs),
-            geometry=self.geometry.name,
-            crs=self.crs,
-        )
 
     def to_postgis(
         self,
@@ -2159,7 +2386,7 @@ individually so that features may have different properties
         Upload GeoDataFrame into PostGIS database.
 
         This method requires SQLAlchemy and GeoAlchemy2, and a PostgreSQL
-        Python driver (e.g. psycopg2) to be installed.
+        Python driver (psycopg or psycopg2) to be installed.
 
         It is also possible to use :meth:`~GeoDataFrame.to_file` to write to a database.
         Especially for file geodatabases like GeoPackage or SpatiaLite this can be
@@ -2212,48 +2439,7 @@ individually so that features may have different properties
             self, name, con, schema, if_exists, index, index_label, chunksize, dtype
         )
 
-        #
-        # Implement standard operators for GeoSeries
-        #
-
-    def __xor__(self, other):
-        """Implement ^ operator as for builtin set type"""
-        warnings.warn(
-            "'^' operator will be deprecated. Use the 'symmetric_difference' "
-            "method instead.",
-            FutureWarning,
-            stacklevel=2,
-        )
-        return self.geometry.symmetric_difference(other)
-
-    def __or__(self, other):
-        """Implement | operator as for builtin set type"""
-        warnings.warn(
-            "'|' operator will be deprecated. Use the 'union' method instead.",
-            FutureWarning,
-            stacklevel=2,
-        )
-        return self.geometry.union(other)
-
-    def __and__(self, other):
-        """Implement & operator as for builtin set type"""
-        warnings.warn(
-            "'&' operator will be deprecated. Use the 'intersection' method instead.",
-            FutureWarning,
-            stacklevel=2,
-        )
-        return self.geometry.intersection(other)
-
-    def __sub__(self, other):
-        """Implement - operator as for builtin set type"""
-        warnings.warn(
-            "'-' operator will be deprecated. Use the 'difference' method instead.",
-            FutureWarning,
-            stacklevel=2,
-        )
-        return self.geometry.difference(other)
-
-    plot = CachedAccessor("plot", geopandas.plotting.GeoplotAccessor)
+    plot = Accessor("plot", geopandas.plotting.GeoplotAccessor)
 
     @doc(_explore)
     def explore(self, *args, **kwargs) -> folium.Map:
@@ -2292,6 +2478,16 @@ individually so that features may have different properties
             Suffix to apply to overlapping column names (left GeoDataFrame).
         rsuffix : string, default 'right'
             Suffix to apply to overlapping column names (right GeoDataFrame).
+        distance : number or array_like, optional
+            Distance(s) around each input geometry within which to query the tree
+            for the 'dwithin' predicate. If array_like, must be
+            one-dimesional with length equal to length of left GeoDataFrame.
+            Required if ``predicate='dwithin'``.
+        on_attribute : string, list or tuple
+            Column name(s) to join on as an additional join restriction on top
+            of the spatial predicate. These must be found in both DataFrames.
+            If set, observations are joined only if the predicate applies
+            and values in specified columns match.
 
         Examples
         --------
@@ -2314,22 +2510,22 @@ individually so that features may have different properties
         [5 rows x 9 columns]
 
         >>> groceries.head()  # doctest: +SKIP
-           OBJECTID     Ycoord  ...  Category                         geometry
-        0        16  41.973266  ...       NaN  MULTIPOINT (-87.65661 41.97321)
-        1        18  41.696367  ...       NaN  MULTIPOINT (-87.68136 41.69713)
-        2        22  41.868634  ...       NaN  MULTIPOINT (-87.63918 41.86847)
-        3        23  41.877590  ...       new  MULTIPOINT (-87.65495 41.87783)
-        4        27  41.737696  ...       NaN  MULTIPOINT (-87.62715 41.73623)
+           OBJECTID     Ycoord  ...  Category                           geometry
+        0        16  41.973266  ...       NaN  MULTIPOINT ((-87.65661 41.97321))
+        1        18  41.696367  ...       NaN  MULTIPOINT ((-87.68136 41.69713))
+        2        22  41.868634  ...       NaN  MULTIPOINT ((-87.63918 41.86847))
+        3        23  41.877590  ...       new  MULTIPOINT ((-87.65495 41.87783))
+        4        27  41.737696  ...       NaN  MULTIPOINT ((-87.62715 41.73623))
         [5 rows x 8 columns]
 
         >>> groceries_w_communities = groceries.sjoin(chicago)
         >>> groceries_w_communities[["OBJECTID", "community", "geometry"]].head()
-             OBJECTID    community                         geometry
-        0          16       UPTOWN  MULTIPOINT (-87.65661 41.97321)
-        87        365       UPTOWN  MULTIPOINT (-87.65465 41.96138)
-        90        373       UPTOWN  MULTIPOINT (-87.65598 41.96297)
-        140       582       UPTOWN  MULTIPOINT (-87.67417 41.96977)
-        1          18  MORGAN PARK  MULTIPOINT (-87.68136 41.69713)
+           OBJECTID       community                           geometry
+        0        16          UPTOWN  MULTIPOINT ((-87.65661 41.97321))
+        1        18     MORGAN PARK  MULTIPOINT ((-87.68136 41.69713))
+        2        22  NEAR WEST SIDE  MULTIPOINT ((-87.63918 41.86847))
+        3        23  NEAR WEST SIDE  MULTIPOINT ((-87.65495 41.87783))
+        4        27         CHATHAM  MULTIPOINT ((-87.62715 41.73623))
 
         Notes
         -----
@@ -2413,7 +2609,7 @@ individually so that features may have different properties
         ... ).to_crs(groceries.crs)
 
         >>> chicago.head()  # doctest: +SKIP
-            ComAreaID  ...                                           geometry
+           ComAreaID  ...                                           geometry
         0         35  ...  POLYGON ((-87.60914 41.84469, -87.60915 41.844...
         1         36  ...  POLYGON ((-87.59215 41.81693, -87.59231 41.816...
         2         37  ...  POLYGON ((-87.62880 41.80189, -87.62879 41.801...
@@ -2422,19 +2618,19 @@ individually so that features may have different properties
         [5 rows x 87 columns]
 
         >>> groceries.head()  # doctest: +SKIP
-            OBJECTID     Ycoord  ...  Category                         geometry
-        0        16  41.973266  ...       NaN  MULTIPOINT (-87.65661 41.97321)
-        1        18  41.696367  ...       NaN  MULTIPOINT (-87.68136 41.69713)
-        2        22  41.868634  ...       NaN  MULTIPOINT (-87.63918 41.86847)
-        3        23  41.877590  ...       new  MULTIPOINT (-87.65495 41.87783)
-        4        27  41.737696  ...       NaN  MULTIPOINT (-87.62715 41.73623)
+           OBJECTID     Ycoord  ...  Category                           geometry
+        0        16  41.973266  ...       NaN  MULTIPOINT ((-87.65661 41.97321))
+        1        18  41.696367  ...       NaN  MULTIPOINT ((-87.68136 41.69713))
+        2        22  41.868634  ...       NaN  MULTIPOINT ((-87.63918 41.86847))
+        3        23  41.877590  ...       new  MULTIPOINT ((-87.65495 41.87783))
+        4        27  41.737696  ...       NaN  MULTIPOINT ((-87.62715 41.73623))
         [5 rows x 8 columns]
 
         >>> groceries_w_communities = groceries.sjoin_nearest(chicago)
         >>> groceries_w_communities[["Chain", "community", "geometry"]].head(2)
-                     Chain community                              geometry
-        0   VIET HOA PLAZA    UPTOWN  MULTIPOINT (1168268.672 1933554.350)
-        87      JEWEL OSCO    UPTOWN  MULTIPOINT (1168837.980 1929246.962)
+                       Chain    community                                geometry
+        0     VIET HOA PLAZA       UPTOWN   MULTIPOINT ((1168268.672 1933554.35))
+        1  COUNTY FAIR FOODS  MORGAN PARK  MULTIPOINT ((1162302.618 1832900.224))
 
 
         To include the distances:
@@ -2442,10 +2638,10 @@ individually so that features may have different properties
         >>> groceries_w_communities = groceries.sjoin_nearest(chicago, \
 distance_col="distances")
         >>> groceries_w_communities[["Chain", "community", \
-"distances"]].head(2)  # doctest: +SKIP
-                     Chain community  distances
-        0   VIET HOA PLAZA    UPTOWN        0.0
-        87      JEWEL OSCO    UPTOWN        0.0
+"distances"]].head(2)
+                       Chain    community  distances
+        0     VIET HOA PLAZA       UPTOWN        0.0
+        1  COUNTY FAIR FOODS  MORGAN PARK        0.0
 
         In the following example, we get multiple groceries for Uptown because all
         results are equidistant (in this case zero because they intersect).
@@ -2455,7 +2651,7 @@ distance_col="distances")
 distance_col="distances", how="right")
         >>> uptown_results = \
 chicago_w_groceries[chicago_w_groceries["community"] == "UPTOWN"]
-        >>> uptown_results[["Chain", "community"]]  # doctest: +SKIP
+        >>> uptown_results[["Chain", "community"]]
                     Chain community
         30  VIET HOA PLAZA    UPTOWN
         30      JEWEL OSCO    UPTOWN
@@ -2486,7 +2682,7 @@ chicago_w_groceries[chicago_w_groceries["community"] == "UPTOWN"]
             exclusive=exclusive,
         )
 
-    def clip(self, mask, keep_geom_type: bool = False) -> GeoDataFrame:
+    def clip(self, mask, keep_geom_type: bool = False, sort=False) -> GeoDataFrame:
         """Clip points, lines, or polygon geometries to the mask extent.
 
         Both layers must be in the same Coordinate Reference System (CRS).
@@ -2509,6 +2705,10 @@ chicago_w_groceries[chicago_w_groceries["community"] == "UPTOWN"]
             If True, return only geometries of original type in case of intersection
             resulting in multiple geometry types or GeometryCollections.
             If False, return all resulting geometries (potentially mixed types).
+        sort : boolean, default False
+            If True, the order of rows in the clipped GeoDataFrame will be preserved at
+            small performance cost. If False the order of rows in the clipped
+            GeoDataFrame will be random.
 
         Returns
         -------
@@ -2539,7 +2739,7 @@ chicago_w_groceries[chicago_w_groceries["community"] == "UPTOWN"]
         >>> nws_groceries.shape
         (7, 8)
         """
-        return geopandas.clip(self, mask=mask, keep_geom_type=keep_geom_type)
+        return geopandas.clip(self, mask=mask, keep_geom_type=keep_geom_type, sort=sort)
 
     def overlay(
         self,
@@ -2572,7 +2772,7 @@ chicago_w_groceries[chicago_w_groceries["community"] == "UPTOWN"]
             geometries.
         make_valid : bool, default True
             If True, any invalid input geometries are corrected with a call to
-            `buffer(0)`, if False, a `ValueError` is raised if any input geometries
+            make_valid(), if False, a `ValueError` is raised if any input geometries
             are invalid.
 
         Returns
@@ -2592,40 +2792,40 @@ chicago_w_groceries[chicago_w_groceries["community"] == "UPTOWN"]
         >>> df2 = geopandas.GeoDataFrame({'geometry': polys2, 'df2_data':[1,2]})
 
         >>> df1.overlay(df2, how='union')
-        df1_data  df2_data                                           geometry
-        0       1.0       1.0  POLYGON ((2.00000 2.00000, 2.00000 1.00000, 1....
-        1       2.0       1.0  POLYGON ((2.00000 2.00000, 2.00000 3.00000, 3....
-        2       2.0       2.0  POLYGON ((4.00000 4.00000, 4.00000 3.00000, 3....
-        3       1.0       NaN  POLYGON ((2.00000 0.00000, 0.00000 0.00000, 0....
-        4       2.0       NaN  MULTIPOLYGON (((3.00000 4.00000, 3.00000 3.000...
-        5       NaN       1.0  MULTIPOLYGON (((2.00000 3.00000, 2.00000 2.000...
-        6       NaN       2.0  POLYGON ((3.00000 5.00000, 5.00000 5.00000, 5....
+           df1_data  df2_data                                           geometry
+        0       1.0       1.0                POLYGON ((2 2, 2 1, 1 1, 1 2, 2 2))
+        1       2.0       1.0                POLYGON ((2 2, 2 3, 3 3, 3 2, 2 2))
+        2       2.0       2.0                POLYGON ((4 4, 4 3, 3 3, 3 4, 4 4))
+        3       1.0       NaN      POLYGON ((2 0, 0 0, 0 2, 1 2, 1 1, 2 1, 2 0))
+        4       2.0       NaN  MULTIPOLYGON (((3 4, 3 3, 2 3, 2 4, 3 4)), ((4...
+        5       NaN       1.0  MULTIPOLYGON (((2 3, 2 2, 1 2, 1 3, 2 3)), ((3...
+        6       NaN       2.0      POLYGON ((3 5, 5 5, 5 3, 4 3, 4 4, 3 4, 3 5))
 
         >>> df1.overlay(df2, how='intersection')
-        df1_data  df2_data                                           geometry
-        0         1         1  POLYGON ((2.00000 2.00000, 2.00000 1.00000, 1....
-        1         2         1  POLYGON ((2.00000 2.00000, 2.00000 3.00000, 3....
-        2         2         2  POLYGON ((4.00000 4.00000, 4.00000 3.00000, 3....
+           df1_data  df2_data                             geometry
+        0         1         1  POLYGON ((2 2, 2 1, 1 1, 1 2, 2 2))
+        1         2         1  POLYGON ((2 2, 2 3, 3 3, 3 2, 2 2))
+        2         2         2  POLYGON ((4 4, 4 3, 3 3, 3 4, 4 4))
 
         >>> df1.overlay(df2, how='symmetric_difference')
-        df1_data  df2_data                                           geometry
-        0       1.0       NaN  POLYGON ((2.00000 0.00000, 0.00000 0.00000, 0....
-        1       2.0       NaN  MULTIPOLYGON (((3.00000 4.00000, 3.00000 3.000...
-        2       NaN       1.0  MULTIPOLYGON (((2.00000 3.00000, 2.00000 2.000...
-        3       NaN       2.0  POLYGON ((3.00000 5.00000, 5.00000 5.00000, 5....
+           df1_data  df2_data                                           geometry
+        0       1.0       NaN      POLYGON ((2 0, 0 0, 0 2, 1 2, 1 1, 2 1, 2 0))
+        1       2.0       NaN  MULTIPOLYGON (((3 4, 3 3, 2 3, 2 4, 3 4)), ((4...
+        2       NaN       1.0  MULTIPOLYGON (((2 3, 2 2, 1 2, 1 3, 2 3)), ((3...
+        3       NaN       2.0      POLYGON ((3 5, 5 5, 5 3, 4 3, 4 4, 3 4, 3 5))
 
         >>> df1.overlay(df2, how='difference')
-                                                geometry  df1_data
-        0  POLYGON ((2.00000 0.00000, 0.00000 0.00000, 0....         1
-        1  MULTIPOLYGON (((3.00000 4.00000, 3.00000 3.000...         2
+                                                    geometry  df1_data
+        0      POLYGON ((2 0, 0 0, 0 2, 1 2, 1 1, 2 1, 2 0))         1
+        1  MULTIPOLYGON (((3 4, 3 3, 2 3, 2 4, 3 4)), ((4...         2
 
         >>> df1.overlay(df2, how='identity')
-        df1_data  df2_data                                           geometry
-        0       1.0       1.0  POLYGON ((2.00000 2.00000, 2.00000 1.00000, 1....
-        1       2.0       1.0  POLYGON ((2.00000 2.00000, 2.00000 3.00000, 3....
-        2       2.0       2.0  POLYGON ((4.00000 4.00000, 4.00000 3.00000, 3....
-        3       1.0       NaN  POLYGON ((2.00000 0.00000, 0.00000 0.00000, 0....
-        4       2.0       NaN  MULTIPOLYGON (((3.00000 4.00000, 3.00000 3.000...
+           df1_data  df2_data                                           geometry
+        0       1.0       1.0                POLYGON ((2 2, 2 1, 1 1, 1 2, 2 2))
+        1       2.0       1.0                POLYGON ((2 2, 2 3, 3 3, 3 2, 2 2))
+        2       2.0       2.0                POLYGON ((4 4, 4 3, 3 3, 3 4, 4 4))
+        3       1.0       NaN      POLYGON ((2 0, 0 0, 0 2, 1 2, 1 1, 2 1, 2 0))
+        4       2.0       NaN  MULTIPOLYGON (((3 4, 3 3, 2 3, 2 4, 3 4)), ((4...
 
         See also
         --------
@@ -2645,7 +2845,7 @@ chicago_w_groceries[chicago_w_groceries["community"] == "UPTOWN"]
 def _dataframe_set_geometry(
     self,
     col,
-    drop: bool = False,
+    drop: bool | None = None,
     inplace: Literal[False] = False,
     crs: Any | None = None,
 ) -> GeoDataFrame:
