@@ -5,8 +5,11 @@ import pytest
 
 
 @pytest.mark.parametrize(
-    "test_dataset", ["naturalearth_lowres", "naturalearth_cities", "nybb"]
+    "test_dataset", ["naturalearth_lowres", "naturalearth_cities", "nybb", "foo"]
 )
 def test_read_paths(test_dataset):
-    with pytest.warns(FutureWarning, match="The geopandas.dataset module is"):
+    with pytest.raises(
+        AttributeError,
+        match=r"The geopandas\.dataset has been deprecated and was removed",
+    ):
         assert isinstance(read_file(get_path(test_dataset)), GeoDataFrame)
