@@ -10,7 +10,6 @@ def cumsum(whatever):
 
     It computes the cumulative {operation}.
     """
-    ...
 
 
 @doc(
@@ -27,18 +26,15 @@ def cumsum(whatever):
     method="cumavg",
     operation="average",
 )
-def cumavg(whatever):
-    ...
+def cumavg(whatever): ...
 
 
 @doc(cumsum, method="cummax", operation="maximum")
-def cummax(whatever):
-    ...
+def cummax(whatever): ...
 
 
 @doc(cummax, method="cummin", operation="minimum")
-def cummin(whatever):
-    ...
+def cummin(whatever): ...
 
 
 def test_docstring_formatting():
