@@ -1,8 +1,8 @@
+from .clip import clip
 from .geocoding import geocode, reverse_geocode
 from .overlay import overlay
 from .sjoin import sjoin, sjoin_nearest
 from .util import collect
-from .clip import clip
 
 __all__ = [
     "collect",
