@@ -1,13 +1,14 @@
 from warnings import warn
 
 import numpy
+
 from shapely.geometry import MultiPoint
 
 from geopandas.array import from_shapely, points_from_xy
 from geopandas.geoseries import GeoSeries
 
 
-def uniform(geom, size, seed=None):
+def uniform(geom, size, rng=None):
     """
 
     Sample uniformly at random from a geometry.
@@ -39,7 +40,7 @@ def uniform(geom, size, seed=None):
     >>> square = box(0,0,1,1)
     >>> uniform(square, size=102) # doctest: +SKIP
     """
-    generator = numpy.random.default_rng(seed=seed)
+    generator = numpy.random.default_rng(seed=rng)
 
     if geom is None or geom.is_empty:
         return MultiPoint()
@@ -64,7 +65,7 @@ def _uniform_line(geom, size, generator):
     """
 
     fracs = generator.uniform(size=size)
-    return from_shapely(geom.interpolate(fracs, normalized=True)).unary_union()
+    return from_shapely(geom.interpolate(fracs, normalized=True)).union_all()
 
 
 def _uniform_polygon(geom, size, generator):
@@ -80,4 +81,4 @@ def _uniform_polygon(geom, size, generator):
         )
         valid_samples = batch[batch.sindex.query(geom, predicate="contains")]
         candidates.extend(valid_samples)
-    return GeoSeries(candidates[:size]).unary_union
+    return GeoSeries(candidates[:size]).union_all()
